### 2020-05-22 #2221 @lskatz
Thank you @arfon!  This is the second manuscript that I have put through JOSS and the first one, [Mashtree](https://github.com/openjournals/joss-reviews/issues/1750), was a good experience.  Additionally if it helps makes things go more smoothly, I would like to highlight that I signed up as a reviewer last year and would be more than glad to help review for JOSS.

I think that some potential reviewers could be:
* tseemann
* erinyoung
* cimendes

I also went through your spreadsheet and found some reviewers that are familiar with perl and bioinformatics:
* tkchafin
* bede
* CFGrote
* druvus
