+++ scripts/sn_createSampleSheet.pl
@@ -1,0 +1,105 @@
+#!/usr/bin/env perl
+# Creates a sample sheet for a folder
+
+use strict;
+use warnings;
+use Getopt::Long;
+use Data::Dumper;
+use File::Copy qw/mv/;
+use File::Path qw/remove_tree/;
+use File::Basename qw/fileparse basename dirname/;
+use File::Temp qw/tempdir/;
+use File::Find qw/find/;
+use FindBin;
+
+use lib "$FindBin::RealBin/../lib/perl5";
+use SneakerNet qw/readConfig command logmsg/;
+
+$ENV{PATH}="$ENV{PATH}:/opt/cg_pipeline/scripts";
+
+local $0=fileparse $0;
+
+exit(main());
+
+sub main{
+  my $settings={};
+  GetOptions($settings,qw(help tempdir=s numcpus=i )) or die $!;
+  $$settings{tempdir}||=tempdir("$0.XXXXXX",TMPDIR=>1,CLEANUP=>1);
+  mkdir($$settings{tempdir}) if(!-e $$settings{tempdir});
+  $$settings{numcpus}||=1;
+
+  die usage() if($$settings{help} || !@ARGV);
+
+  my $samplesheet = "$ARGV[0]/samples.tsv";
+  createSampleSheetOutOfThinAir($samplesheet, $ARGV[0], $settings);
+
+  return 0;
+}
+
+# Using lots of help from Taylor Griswold for creating the sample sheet
+# out of thin air.
+sub createSampleSheetOutOfThinAir{
+  my($outSamplesheet, $indir, $settings) =@_;
+
+  # get a list of fastq files for the next couple of steps
+  # Do not include the full path; just the basename.
+  #my @fastq = map{basename($_)} glob("$indir/*.fastq.gz");
+  my %fastq; # basename => full path
+  find({wanted=>sub{
+    my $path = $File::Find::name;
+    return if(!-f $path);
+    return if($path !~ /\.fastq.gz$/);
+
+    my $basename = basename($path);
+
+    $fastq{$basename} = $path;
+  }}, $indir);
+  my @fastq = sort {$a cmp $b} keys(%fastq);
+  
+  #awk -F'_' '{print $1}' <(ls *.fastq.gz) | sed 's/-*$//g' | uniq > wgs-ids_detailed.txt 
+  my %sample;
+  for my $fastq(@fastq){
+    my($sample) = split(/_/, $fastq);
+    $sample =~ s/\-*$//; # remove trailing dashes
+    $sample{$sample}=1;
+  }
+  my @sample = sort {$a cmp $b} keys(%sample);  # sample list is unique and sorted
+
+  # for i in $(cat wgs-ids_detailed.txt); do R1=`realpath $(ls *$i*_R1*.fastq.gz)`; R2=`realpath $(ls *$i*_R2*.fastq.gz)`; wgsid=`basename $R1 ".fastq.gz" | cut -d '_' -f 1`; echo -e $wgsid"\ttaxon=Salmonella;route=calcengine\t"$R1";"$R2; done > M1234-19-004.tsv 
+  open(my $outFh, ">", $outSamplesheet) or die "ERROR: could not write to $outSamplesheet: $!";
+  for my $name(@sample){
+    my $R1 = (grep { $_=~/$name.*_R1/; } @fastq)[0];
+    my $R2 = $R1;
+    $R2 =~ s/_R1/_R2/;
+
+    # Check that the keys exist
+    if(!$fastq{$R1} && !$fastq{$R2}){
+      die "ERROR: R1 and R2 are not both represented as $R1 and $R2";
+    }
+    # Check that the fastq files exist
+    if(!-e $fastq{$R1}){
[…]
