### 2020-06-08 #2221 @lskatz
Thank you @lpantano! Please let me know if there is anything else we can do to facilitate the peer review process.

Two more potential reviewers came to mind and so I just wanted to add onto my earlier comment.

> I think that some potential reviewers could be:
> tseemann
> erinyoung
> cimendes
> k-florek
> HeatherBlankenship
>
> I also went through your spreadsheet and found some reviewers that are familiar with perl and bioinformatics:
> tkchafin
> bede
> CFGrote
> druvus
