### 2020-06-08 #2221 @arfon
@lpantano invites like this only work for editors (not reviewers)
