### 2020-06-11 db90262
changing default assembly to shovill
