+++ SneakerNet.plugins/assembleAll.pl
@@ -1,5 +1,5 @@
 #!/usr/bin/env perl
-# Transfers files to a remote destination and QCs them beforehand.
+# Assemble genomes for Cryptosporidum
 
 use strict;
 use warnings;
@@ -18,8 +18,8 @@ use FindBin;
 use lib "$FindBin::RealBin/../lib/perl5";
 use SneakerNet qw/exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg fullPathToExec/;
 
-our $VERSION = "1.3";
-our $CITATION = "assembleAll.pl by Lee Katz. Uses Skesa and Prodigal for assembly and gene prediction.";
+our $VERSION = "2.0";
+our $CITATION= "Assembly plugin by Lee Katz. Uses SHOvill.";
 
 local $0=fileparse $0;
 exit(main());
@@ -30,36 +30,58 @@ sub main{
   exitOnSomeSneakernetOptions({
       _CITATION => $CITATION,
       _VERSION  => $VERSION,
-      'run_assembly_filterContigs.pl (CG-Pipeline)' => "echo CG Pipeline version unknown",
       'run_prediction_metrics.pl (CG-Pipeline)'     => "echo CG Pipeline version unknown",
-      'skesa'                         => 'skesa --version 2>&1 | grep SKESA',
-      'prodigal'                      => "prodigal -v 2>&1 | grep -i '^Prodigal V'",
+      'run_assembly_metrics.pl (CG-Pipeline)'       => "echo CG Pipeline version unknown",
+      cat                             => 'cat --version | head -n 1',
+      sort                            => 'sort --version | head -n 1',
+      head                            => 'head --version | head -n 1',
+      uniq                            => 'uniq --version | head -n 1',
+      touch                           => 'touch --version | head -n 1',
+      shovill                         => 'shovill --version',
+      prodigal                        => "prodigal -v 2>&1 | grep -i '^Prodigal V'",
+
+      # shovill requires a ton of things:
+      'seqtk'       => 'seqtk 2>&1 | grep Version',
+      'pigz'        => 'pigz --version 2>&1',
+      'mash'        => 'mash --version 2>&1',
+      'trimmomatic' => 'trimmomatic -version 2>&1 | grep -v _JAVA',
+      'lighter'     => 'lighter -v 2>&1',
+      'flash'       => 'flash --version 2>&1 | grep FLASH',
+      'spades.py'   => 'spades.py  --version 2>&1',
+      'skesa'       => 'skesa --version 2>&1 | grep SKESA',
+      'bwa'         => 'bwa 2>&1 | grep Version:',
+      'samtools'    => 'samtools 2>&1 | grep Version:',
+      'samclip'     => 'samclip --version 2>&1',
+      'java'        => 'java -version 2>&1 | grep version',
+      'pilon'       => 'pilon --version 2>&1 | grep -v _JAVA',
+
+      # not using megahit or velvet in this instance of shovill
+      'megahit'     => 'megahit --version 2>&1',
+      'megahit_toolkit' => 'megahit_toolkit dumpversion 2>&1',
+      'velveth'     => 'velveth 2>&1 | grep Version',
+      'velvetg'     => 'velvetg 2>&1 | grep Version',
     }, $settings,
   );
 
-  usage() if($$settings{help} || !@ARGV);
+  die usage() if($$settings{help} || !@ARGV);
[…]
