+++ docs/plugins/assembleAll.pl.md
@@ -7,7 +7,23 @@ predicts genes.
 
 # Software requirements
 
-* Skesa
+In version 2, Shovill was added and many requirements were added
+
+* Shovill - has many dependencies itself
+  * BWA
+  * Flash
+  * Java
+  * Lighter
+  * Mash
+  * Megahit (not used but checked when shovill runs)
+  * pigz
+  * Pilon
+  * Samclip
+  * Samtools
+  * Seqtk
+  * Skesa
+  * Trimmomatic
+  * Velvet  (not used but checked when shovill runs)
 * CG-Pipeline
 * Prodigal
 
