### 2020-06-12 c8c6c19
unit testing good for assembleAll
