### 2020-06-12 #2334 @whedon
[REVIEW]: SneakerNet: A modular quality assurance and quality check workflow for primary genomic and metagenomic read data
**Submitting author:** @lskatz (<a href="http://orcid.org/0000-0002-2533-9161">Lee Katz</a>)
**Repository:** <a href="https://github.com/lskatz/SneakerNet" target ="_blank">https://github.com/lskatz/SneakerNet</a>
**Version:** v0.19.7

@lfaller & @ @erinyoung, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
