### 2020-06-16 lskatz/SneakerNet#38 @lfaller
Please complete the Apache license copyright field
Hello! I am a reviewer for [your tool](https://github.com/openjournals/joss-reviews/issues/2334) in the JOSS 👋

I noticed that you forgot to fill in the copyright line in your Apache license text file:

https://github.com/lskatz/SneakerNet/blob/e590634435fea3c6437d4061d66c3f259579fa48/LICENSE#L189

Would be great if you could fill in that information.

Cheers,
~Lina
