### 2020-06-16 lskatz/SneakerNet#38 @lskatz
Thank you for pointing that out!  Fixed in 2d4e3eea29a26c60c90c6fc75af517cc7b1826eb.
