### 2020-06-20 d553e9e
attempting graceful die when no kraken results are found
