### 2020-06-23 #2334 @erinyoung
@lpantano <https://github.com/lpantano> et al.,

I went to the invitation URL (
https://github.com/openjournals/joss-reviews/invitations), and didn't see
an invitation associated with my github account, or any related
notifications.

Could this be due to my handle being listed as "@ @erinyoung" instead of
"@erinyoung"? (I noticed that my handle has an extra "@" in this thread)
