### 2020-07-07 lskatz/SneakerNet#41 @lfaller
Please consider fully moving to docker for installation purposes
Hello!

I have some feedback regarding the installation process.

I was trying to follow the "quick installation" instructions outlined [here](https://github.com/lskatz/SneakerNet/blob/master/docs/INSTALL.md).

I started with a clean ubuntu box from docker (`docker run -it ubuntu bash`) and after an hour of trying to get the dependencies set up I gave up.

I would recommend you move the installation process fully to docker to enable users of your software to get started quicker.

I appreciate that you provide information regarding using docker [here](https://github.com/lskatz/SneakerNet/blob/master/docs/CONTAINERS.md). However, I would suggest you provide more detail around the example data to get the user started up quicker.

Right now, the document lists obscure input directories:

```
# this assumes that your miseq run directory, and desired SneakerNet input directory are in your PWD.
# you will need to adjust PATHs depending on where your data is located
export MISEQ=my/miseq/run/dir
export INDIR=my/sneakernet/run/dir
```

It is not clear what should be in these directories unless you are familiar with a miseq run directory. This can be difficult for someone who is not as familiar with Illumina data generation. Similarly, what does it mean to have a SneakerNet input directory? It is called "input directory" -- what should that input be? Sequences? Databases? etc Also, where will the output be stored?

Ideally, example data should be bundled with the docker image so that the user doesn't need to download extra data (maybe someone wants to try your software but doesn't have ready access to Illumina miseq data). Then the user can get started with four lines of code:

```
# get the docker image
docker pull lskatz/sneakernet:latest

# test the software
docker run --rm ... test

# run with example data
docker run --rm ... example

# run with your own data
docker run --rm -v ... my_data
```

By the way, I saw that you have more documentation in the `docs/` subdirectory. This is a great resource, but it requires the user to do some digging to get started. I would argue for simplifying this information to make it more easily digestible.

Finally, I would move the above 4 line docker-based installation to the front page of your github repo. I imagine someone interested in using the software would find information about the installation most useful, and the first thing they will look at is the general README of the github repo. With so much open source software out there, software that is easy to install and use will get the most traction.

Anyways, I hope you find this feedback useful.

Related Issue: https://github.com/openjournals/joss-reviews/issues/2334
