### 2020-07-07 lskatz/SneakerNet#42 @lfaller
Consider adding information about compute resources needed
I noticed that the docker image is over 3 Gb when downloaded from dockerhub.

I know some of the tools that SneakerNet uses ship with their own databases. Is that the reason for the size of the docker image?

In order to pare down the docker image, you could consider if these databases are all needed, or if some of them could be optional for the user. You can also look at [multi-stage builds](https://docs.docker.com/develop/develop-images/multistage-build/) and [docker buildkit](https://docs.docker.com/develop/develop-images/build_enhancements/) to try and optimize your docker image.

In general, it would be useful to see an overview of compute resources required to run a basic out-of-the-box SneakerNet analysis.

Related Issue: https://github.com/openjournals/joss-reviews/issues/2334
