### 2020-07-10 c70756e
updated CONTAINERS.md - Docker section [ci skip]
added clear instructions on running SneakerNet on the packaged example data. Still need to adjust and test section on Singularity.
