### 2020-07-10 4db6dee
updated CONTAINERS.md - singularity section [ci skip]
added section for running example data through SneakerNet using singularity
