### 2020-07-10 1b00806
Fixed path to test data; added link to containers
New link to containers under the installation section
