### 2020-07-10 5cc1991
update CONTAINERS.md - singularity [ci skip]
 - Adjusted link to dockerhub to prevent redirect issues in the future.
 - Cleaned up Singularity instructions, added section on running example data through SN
 - added table of contents with links
