### 2020-07-29 d6419da
explanation of emoji output
