### 2020-08-03 #2334 @lskatz
Hi @lpantano and also @erinyoung and @lfaller --
Sorry for my delay.  I had a sort of surprise temporary reassignment for COVID-19 at CDC and was not able to address your excellent feedback.  We (co-authors and I) should be able to address all outstanding comments in a week or two.  Thank you for your patience.
