### 2020-08-03 #2334 @erinyoung
@lskatz ,

I'm actually still having issues downloading sneakernet.

I second the opinion to move everything to docker (https://github.com/lskatz/SneakerNet/issues/41)

There are too many hidden dependencies for the source install. I tried `make test`, and first encountered that I was missing `Config/Simple.pm`, then `Bio/FeatureIO/gff.pm`, then it couldn't find the path to mlst and a few other tools, then I had the wrong version of perl installed (v5.26.0 needed, v5.16.3 installed).

I gave up on the source installation, and moved to containers, but your containers are very large. I've been having to get creative to install it with both singularity and docker.

For singularity, I don't have a large `/scratch` directory in my setup for compressed temp downloads, and squashfs complained when I attempted to download the container with singularity. I re-attempted the download with the "--sandbox" option (`singularity build --sandbox sneakernet.simg docker://lskatz/sneakernet:latest`). I haven't had the time to test this installation, yet.

I'm sure docker has a similar command for large containers, and I am looking on implementing that.
