### 2020-08-03 59aabf9
Update PLUGINSDEV.md
