### 2020-08-05 fa96bf9
moved kraken outside of the container; addressed some reviewer issues
