+++ Dockerfile
@@ -3,7 +3,7 @@ FROM ubuntu:xenial
 LABEL base.image="ubuntu:xenial"
 LABEL container.version="1"
 LABEL software="SneakerNet"
-LABEL software.version="0.10.4"
+LABEL software.version="0.11.0"
 LABEL description="QA/QC pipeline for a MiSeq/HiSeq/Ion Torrent run"
 LABEL website="https://github.com/lskatz/SneakerNet"
 LABEL license="https://github.com/lskatz/SneakerNet/blob/master/LICENSE"
@@ -253,13 +253,13 @@ RUN pip3 install poetry && \
 # python deps (installed via pip3 cmd below) numpy scipy biopython plotly SPARQLWrapper
 RUN pip3 install chewbbaca==2.1.0
 
-# Get SneakerNet 0.10.4 and make /data
+# Get SneakerNet 0.11.0 and make /data
 # apt deps: sendmail-base zip bsdmainutils (for column command)
 # perl modules listed in cpanm comments above (some installed there, remaining installed w cpanm command below)
 # 
 # The command at the end t/00_env.t sets up an environment including generating contaminated reads
 # for further unit testing.
-ENV SNVER=0.10.4
+ENV SNVER=0.11.0
 RUN wget https://github.com/lskatz/SneakerNet/archive/v${SNVER}.tar.gz && \
  tar -zxf v${SNVER}.tar.gz && \
  rm v${SNVER}.tar.gz && \
@@ -267,16 +267,16 @@ RUN wget https://github.com/lskatz/SneakerNet/archive/v${SNVER}.tar.gz && \
  cpanm --installdeps --notest --force . && \
  perl Makefile.PL && \
  make && \
- sed -i 's+/opt/kraken/full-20140723+/kraken-database/minikraken_20171013_4GB+g' config/settings.conf && \
- mkdir /data && \
+ sed -i 's+/opt/kraken/full-20140723+/kraken-database+g' config/settings.conf && \
+ mkdir /data /kraken-database && \
  perl t/00_env.t
 
 # minikraken db
-RUN mkdir /kraken-database && \
- cd /kraken-database && \
- wget  https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz && \
- tar -zxf minikraken_20171019_4GB.tgz && \
- rm -rf minikraken_20171019_4GB.tgz 
+#RUN mkdir /kraken-database && \
+# cd /kraken-database && \
+# wget  https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz && \
+# tar -zxf minikraken_20171019_4GB.tgz && \
+# rm -rf minikraken_20171019_4GB.tgz 
 
 # set PATH and perl local settings
 ENV PATH="${PATH}:\
