### 2020-08-05 3f83d27
added a mention to the test data in CONTAINERS.md
