### 2020-08-06 #2334 @lskatz
Hi @erinyoung , we are in process of a larger reply to the comments above from both you and @lfaller .  However for right now one solution might be to unset $SINGULARITY_TMPDIR so that the system default takes hold.

@erinyoung and @lfaller For the latest container (again, we are making a much larger reply to address the comments in full), we removed the Kraken database and provided instructions on how to download the Kraken database separately.  The container size changed significantly from 3.9G to 1.1G.  Hopefully these two changes can help you get over those immediate obstacles while we address the other comments.

@kapsakcj is the main expert on the container front and so I am tagging him for any further conversation along these lines.
