### 2020-08-07 14b4104
Sanitizing SampleSheet.csv in case of whitespace characters
