### 2020-08-07 1c37da6
Fixed an issue where a SampleSheet.csv might be altered by Excel on the user's end accidentally
