### 2020-08-07 lskatz/SneakerNet#41 @lskatz
Hi @lfaller and @erinyoung, thank you for these comments.

I have released v0.11.1 which has improvements on Docker and the documentation as suggested:
* better description of the MiSeq directory
* Better quickstart with Docker in the main `README.md`
* `README.md` for the docs subdirectory, to make a catalog of help files
* A link at the bottom of `README.md` to `docs/README.md`
* Pushed some wording from CONTAINERS.md to main README.md
