### 2020-08-07 lskatz/SneakerNet#42 @lskatz
Hi @lfaller and @erinyoung,

I have removed the Kraken database from the container between v0.10 and v0.11 and the size of the container dropped significantly.  There are also instructions in CONTAINERS.md on how to link a Kraken database when running the SneakerNet container.

I am still benchmarking but wanted to give an update before the end of the week.  Thank you for your comments.
