### 2020-08-10 e769008
Fixed quality failing error; added more debug statements with --debug; v5.0
