### 2020-08-10 a37f19b
v0.11.2: fixed passfail bug with quality scores
