### 2020-08-10 lskatz/SneakerNet#42 @lfaller
That sounds great, thanks for the update @lskatz !
