### 2020-08-11 lskatz/SneakerNet#42 @lskatz
For sure, @lfaller !  Thank you for your helpful comments and suggestions!

> In general, it would be useful to see an overview of compute resources required to run a basic out-of-the-box SneakerNet analysis.

I benchmarked SneakerNet on the test data (four genomes at 10x + one contaminated genome which is a mixture of two of the genomes).  These benchmarks were performed on a 72-cpu computer to help avoid any collision of resources in the concurrent test.  Benchmarks were measured with `time --verbose`.

    for i in 1 2 4 8 16 32; do
      cp -r M00123-18-001-test M00123-18-001-test-$i{cpu}
      \time --verbose singularity exec -B $PWD:/data -B $KRAKEN_DEFAULT_DB:/kraken-database sneakernet.0.11.2.sif SneakerNetPlugins.pl --numcpus $i --no email --no transfer --no save --no immediate M00123-18-001-test-${i}cpu > benchmark.${i}cpus.log 2>&1 &
     done;

With four threads (something I would assume to be reasonable for many or most people) and using the container version of v0.11.2, I think the most relevant stats are that it took about 17 minutes to complete and about 6.2G of RAM with 4 cpus.  I am attaching the full benchmarks for review.

[benchmark.1cpus.log](https://github.com/lskatz/SneakerNet/files/5058036/benchmark.1cpus.log)
[benchmark.2cpus.log](https://github.com/lskatz/SneakerNet/files/5058037/benchmark.2cpus.log)
[benchmark.4cpus.log](https://github.com/lskatz/SneakerNet/files/5058038/benchmark.4cpus.log)
[benchmark.8cpus.log](https://github.com/lskatz/SneakerNet/files/5058039/benchmark.8cpus.log)
[benchmark.16cpus.log](https://github.com/lskatz/SneakerNet/files/5058034/benchmark.16cpus.log)
[benchmark.32cpus.log](https://github.com/lskatz/SneakerNet/files/5058035/benchmark.32cpus.log)

I am also extracting the RAM and time per cpu in this graph, where the first Y-axis is maximum RAM used and the second Y-axis is time elapsed.

![image](https://user-images.githubusercontent.com/14798/89922155-a702a080-dbcc-11ea-9ef0-b2f78535d06d.png)
