### 2020-08-11 lskatz/SneakerNet#42 @lskatz
Note: I am running up to 5 or 10 replicates in case it makes the graph any more informative and to help "average out" any outliers, and I should have results soon.
