### 2020-08-12 lskatz/SneakerNet#42 @lskatz
I went big and ran 100 replicates on our HPC.  Please let me know if I addressed all comments on this issue or if there is any followup, which I would be glad to address.

![image](https://user-images.githubusercontent.com/14798/90025706-6d8d6c00-dc84-11ea-8a0d-bd706d59eac5.png)

cpus | mem (kb) | stddev | seconds | stddev
-- | -- | -- | -- | --
1 | 4312189 | 512452.8 | 1215.39 | 231.1566
2 | 4743979 | 859988.6 | 767.87 | 268.9702
4 | 5442763 | 805141.1 | 583.5 | 221.5411
8 | 6114553 | 727898.3 | 466.07 | 158.9255
16 | 6612363 | 490417.6 | 430.88 | 189.8042
32 | 6768286 | 441932.5 | 392.4 | 277.2188

The HPC benchmarking script is below.

```bash
#!/bin/bash
set -e
set -u

for rep in {1..100}; do
  logfolder=rep$rep.qsub;
  [[ ! -d $logfolder ]] && mkdir $logfolder;
  for cpus in 1 2 4 8 16 32; do
    outdir="M00123-18-001-test-${cpus}cpu-${rep}rep"
    logfile="$logfolder/benchmark.${cpus}cpus.log"

    if [[ -e "$logfile" ]]; then

      exit_code=$(grep "Exit status" < $logfile | grep '[0-9]' | true)
      if [[ "$exit_code" == 0 ]]; then
        echo "Found $logfile completed, exit code $exit_code. Skipping.";
        continue
      fi
    fi

    # We are now moving forward because either the logfile
    # is not present or we ran into an error on a node.

    # Delete old logfile
    if [ -e "$logfile" ]; then
      mv -v $logfile $logfile.tmp && rm -v $logfile.tmp &
    fi

    # Submit the job with the appropriate CPUs.
    # Use fast scratch storage.
    qsub -q edlb.q -pe smp $cpus -cwd -V -N "benchmark${cpus}_${rep}" -o $logfolder/benchmark.${cpus}cpus.log -j y -v "KRAKEN_DEFAULT_DB=$KRAKEN_DEFAULT_DB" -v "outdir=$outdir" <<- "END_OF_SCRIPT"
#!/bin/bash
      set -e
      #set -x
      set -u
      which singularity
      hostname
      echo "";

      # Make the folder
      rm -rf $outdir
      outdirAbs=/scratch/$USER/SneakerNetTest/$outdir
      mkdir -pv $(dirname $outdir)
      cp -rv M00123-18-001-test $outdirAbs
      rm -rf $outdirAbs/SneakerNet $outdirAbs/readMetrics.tsv

      tree -L 2 $outdir

      trap ' { rm -rf $outdir; } ' EXIT

      echo ""
      /usr/bin/time --verbose singularity exec -B /scratch:/scratch -B $PWD:/data -B $KRAKEN_DEFAULT_DB:/kraken-database sneakernet.0.11.2.sif SneakerNetPlugins.pl --numcpus $NSLOTS --no email --no transfer --no save --no immediate $outdirAbs

END_OF_SCRIPT
  done;
  #echo "DEBUG"; break;
done
```

This is the perl one-liner I used to extract a table of means and standard deviations from the `time --verbose` output.

```perl -MStatistics::Basic=mean,stddev -MData::Dumper -Mautodie -le '$/=undef; for my $log(glob("rep*/*.log")){open(my $fh, $log); $txt=<$fh>; close $fh; if($log=~/(\d+)\D+(\d+)/){$rep=$1; $cpus=$2;} else{die "ERROR: no cpus or reps in $log filename"} if($txt=~/Elapsed\D+(\d+):(\d+)/){$min=$1; $sec=$2; $time=$min*60+$sec}else{die "ERROR: no time in $log"} if($txt=~/Maximum resident set size\D+(\d+)/){$size=$1} else{die "ERROR: no size in $Log"} $size{$cpus}[$rep-1]=$size; $time{$cpus}[$rep-1]=$time;} for my $cpus(sort{$a<=>$b} keys(%time)){ print join("\t", $cpus, mean(@{$size{$cpus}})+0, stddev(@{$size{$cpus}})+0, mean(@{$time{$cpus}})+0, stddev(@{$time{$cpus}})+0 );}' | column -t```
