### 2020-08-12 lskatz/SneakerNet#42 @lskatz
In retrospect, I have summarized these findings into the documentation here: https://github.com/lskatz/SneakerNet/blob/master/docs/INSTALL.md#resources
