### 2020-08-18 #2334 @lskatz
Hi @erinyoung and @lfaller,
I thought I would check in and bring the conversation back to here (although, it was very useful to have the conversation on the SneakerNet site for posterity and for future users!).  Recently, we updated the container and its instructions to address some of your concerns here in this thread, and in a couple of threads over here: https://github.com/lskatz/SneakerNet/issues/41, https://github.com/lskatz/SneakerNet/issues/42.

In SneakerNet issue 42, there are more details on how the Kraken database was removed from the container and how instructions were put in place for building the new version of container.  Additionally, we added benchmarking information.

In SneakerNet issue 41, we described how we added better documentation.

If you would like to continue testing with the container, please confirm that it is the latest version v0.11.2.  I believe that Dockerhub is not updating the `:latest` tag and so I will look into that.
