### 2020-08-18 7b72f3e
added some compression to colorid log file; addl logmsgs; Clostridum min coverage => 30x per CDC recommendations
