### 2020-08-24 lskatz/SneakerNet#41 @kapsakcj
Hi @lfaller , thanks for the feedback. Do you have access to a computer with more than 1.94GB of RAM? It looks like `shovill` is running out of RAM for each of the samples included in the test data.

Through the benchmarks that @lskatz completed, he found that roughly 4GB minimum of RAM is required to run SneakerNet, and that it it benefits by having more RAM available (recommends 8GB or more)  https://github.com/lskatz/SneakerNet/blob/master/docs/INSTALL.md#resources

----

>Looks like shovill is not happy. Since this is an early step in the pipeline, does that mean nothing will work?

Some of the steps of the SneakerNet pipeline will run with no errors even if `shovill` fails, but some of the steps do require a genome assembly, which is the output of `shovill`. @lskatz has configured SneakerNet to keep progressing even if a step, like `shovill` for example, fails.

I was able to replicate similar, but not exact issues when I limited the docker container to 2GB of RAM. Using this command:
```
docker run -m 2g --rm -v $PWD:/data -v $KRAKEN_DEFAULT_DB:/kraken-database -u $(id -u):$(id -g) lskatz/sneakernet:latest SneakerNetPlugins.pl --numcpus 2 --no email --no transfer --no save /data/$INDIR &> log-2gbRAMlimit.txt`
```
Pilon failed here (one step in the shovill pipeline):
```
...
[shovill] Running: (bwa mem -v 3 -x intractg -t 2 skesa.fasta R1.fq.gz R2.fq.gz | samclip --ref skesa.fasta.fai | samtools sort --threads 1 -m 16384m --reference skesa.fasta -T /tmp/4qpjxLO0My/samtools.119 -o shovill.bam) 2>&1 | sed 's/^/[bwa+samtools-sort] /' | tee -a shovill.log
[shovill] Running: samtools index shovill.bam 2>&1 | sed 's/^/[samtools-index] /' | tee -a shovill.log
[shovill] Correcting errors in skesa.fasta
[shovill] Running: pilon --genome skesa.fasta --frags shovill.bam --minmq 60 --minqual 3 --fix bases --output pilon --threads 2 --changes --mindepth 0.25 2>&1 | sed 's/^/[pilon] /' | tee -a shovill.log
Killed
[shovill] Could not open pilon.changes
assembleAll.pl: shovill failed!
ERROR running command
  shovill --outdir /tmp/assembleAll.pl.X277QU/contaminated --R1 /data/testing_out/contaminated_1.fastq.gz --R2 /data/testing_out/contaminated_2.fastq.gz --assembler skesa --cpus 2 --keepfiles at /SneakerNet-0.11.2/scripts/../SneakerNet.plugins/assembleAll.pl line 288.

assembleAll.pl: ASSEMBLE SAMPLE LT2
...
```

When I allow for more RAM (125GB on my machine, no limit imposed) SneakerNet runs with no errors. I've attached the logfile from this, to give an example of the expected output. I used the same commands that you have listed above:
```
docker run --rm -v $PWD:/data -v $KRAKEN_DEFAULT_DB:/kraken-database -u $(id -u):$(id -g) lskatz/sneakernet:latest SneakerNetPlugins.pl --numcpus 2 --no email --no transfer --no save /data/$INDIR &> log.txt
```
[log.txt](https://github.com/lskatz/SneakerNet/files/5118608/log.txt)

----

Regarding this comment:
> Maybe it indeed failed completely. It didn't generate a report:
>
> ```shell
> $> tail log.txt
> sn_passfail.pl: The pass/fail file is under ./SneakerNet/forEmail/passfail.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/QC_summary.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/passfail.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/staramr.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/basebalance.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/kraken.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/mlst.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/assemblyMetrics.tsv
> sn_report.pl: /data/testing_out/SneakerNet/forEmail/readMetrics.tsv
> sn_report.pl: Report can be found in ./SneakerNet/forEmail/report.html
>
> $> ls -lah ./SneakerNet/forEmail/report.html
> ls: ./SneakerNet/forEmail/report.html: No such file or directory
> ```
SneakerNet isn't aware that it is running inside a docker container, so the output PATH that it prints above ^ is deceiving. Since you specified your `$PWD` on your host system to mount to `/data` inside the container you should find the report within this directory on your host system: `$PWD/testing_out/SneakerNet/forEmail/report.html` after SneakerNet finishes running and the container exits.

Please let us know if you have more questions or encounter any more issues.

Curtis
