### 2020-08-27 ea20eb3
added where assemblies are found
