+++ docs/plugins/assembleAll.pl.md
@@ -35,7 +35,7 @@ then produces assembly and gene prediction metrics.
 
 # Outputs
 
-Assemblies are found in `SneakerNet/assemblies/[samplename]/\*.fasta`
+Assemblies are found in `SneakerNet/assemblies/[samplename]/*.fasta`
 
 Table with columns
 
