### 2020-08-28 d7dd1ee
Available inputs for plugin dev
