### 2020-09-03 7870301
Update PLUGINS.md
