### 2020-09-04 #2334 @erinyoung
Even without the kraken database, it is still a rather large container, so I had to get a little creative on how to install it and get it working.
