### 2020-09-09 #2334 @lskatz
Hi @erinyoung
Good catch!  You are right.  I have made the documentation changes in 09def0a and fc3e6a2.

Thank you for pushing a little harder on the container @erinyoung :)  I am still a novice on creating containers but fortunately we have our expert @kapsakcj.  If we can figure out how to optimize it further, we will.  However, I believe we have an unavoidable problem of many dependencies.
