### 2020-09-11 lskatz/SneakerNet#41 @lfaller
Thanks for the example log, it's helpful to see what I should be expecting!

I increased the resources for Docker to 6 CPUs and 11 GB of RAM.

I am still running into issues with shovill, but I think that's due to not enough RAM:

```
assembleAll.pl: Temporary directory is at /tmp/assembleAll.pl.Hw82B9
assembleAll.pl: ASSEMBLE SAMPLE Philadelphia_CDC
assembleAll.pl: Assembling Philadelphia_CDC
[shovill] Hello stranger
[shovill] You ran: /shovill-1.0.4/bin/shovill --outdir /tmp/assembleAll.pl.Hw82B9/Philadelphia_CDC --R1 /data/testing_out/Philadelphia_CDC_1.fastq.gz --R2 /data/testing_out/Philadelphia_CDC_2.fastq.gz --assembler skesa --cpus 2 --keepfiles
[shovill] This is shovill 1.0.4
[shovill] Written by Torsten Seemann
[shovill] Homepage is https://github.com/tseemann/shovill
[shovill] Operating system is linux
[shovill] Perl version is v5.22.1
[shovill] Machine has 6 CPU cores and 10.72 GB RAM
[shovill] Set --ram to 32 but machine only has 10.72 GB
assembleAll.pl: shovill failed!
ERROR running command
  shovill --outdir /tmp/assembleAll.pl.Hw82B9/Philadelphia_CDC --R1 /data/testing_out/Philadelphia_CDC_1.fastq.gz --R2 /data/testing_out/Philadelphia_CDC_2.fastq.gz --assembler skesa --cpus 2 --keepfiles at /SneakerNet-0.11.2/scripts/../SneakerNet.plugins/assembleAll.pl line 288.
```

I think it would be great to make the memory setting for shovill dynamic if that is possible, but at this point I'd call it a "nice to have". It really does seem to be better to run the pipeline on a high-memory machine 🙂
