### 2020-09-11 lskatz/SneakerNet#42 @lfaller
Thanks for summarizing your benchmark results! I think it'll be helpful for new users to get started to have that kind of information!
