### 2020-09-11 #2334 @lfaller
@lskatz minor suggestion: it would be nice to document what pipeline of plugins Sneakernet runs for the example. I figured it out *after* running the example by looking at the logs, but it would be useful to provide that information in the readme. You could even provide a figure showing that pipeline.
