### 2020-09-14 7bdfb40
Update PLUGINS.md
* Added table of contents
* Added `The exact order of plugins for all workflows can be found by running the command SneakerNet.checkdeps.pl --list.` in the Workflows section
* Added section `Workflows and their plugins` where `SneakerNet.checkdeps.pl --list` is actually documented.
