### 2020-09-14 #2334 @lskatz
@lfaller Great point!  I had a script `SneakerNet.checkdeps.pl` that has this function, but it was not documented well enough.  ([very little documentation here where the script is mentioned once](https://github.com/lskatz/SneakerNet/blob/master/docs/INSTALL.md#software)  and [here where it says cryptically "Workflows are defined in plugins.conf"](https://github.com/lskatz/SneakerNet/blob/master/docs/PLUGINS.md#workflows)

In response, I have changed documentation here: https://github.com/lskatz/SneakerNet/edit/master/docs/PLUGINS.md
* Added table of contents
* Added `The exact order of plugins for all workflows can be found by running the command SneakerNet.checkdeps.pl --list.` in the Workflows section
* Added section `Workflows and their plugins` where `SneakerNet.checkdeps.pl --list` is actually documented.
