### 2020-09-14 1925a69
added default plugins workflow figure
