### 2020-09-14 #2334 @lskatz
@lfaller also in response in the same PLUGINS.md page, I created a basic figure on the default workflow and plugins.
