### 2020-09-15 #2334 @erinyoung
I think I still don't understand how to run the example.

```
> SneakerNet.roRun.pl --createsamplesheet -o M1234-18-001-test /SneakerNet-0.11.2/example/inbox/test-15-001/
A_1.fastq.gz     A_2.fastq.gz     B_1.fastq.gz     B_2.fastq.gz     QC/              SampleSheet.csv  config.xml       readMetrics.tsv
Singularity> SneakerNet.roRun.pl --createsamplesheet -o M1234-18-001-test /SneakerNet-0.11.2/example/inbox/test-15-001
SneakerNet.roRun.pl: demultiplex sample sheets were not found. Creating one directly from fastq filenames into /tmp/SneakerNet.roRun.pl.DaYIcw/runData/samples.tsv
Use of uninitialized value $R2 in substitution (s///) at /SneakerNet-0.11.2/scripts/SneakerNet.roRun.pl line 292.
Use of uninitialized value $R1 in hash element at /SneakerNet-0.11.2/scripts/SneakerNet.roRun.pl line 295.
Use of uninitialized value $R2 in hash element at /SneakerNet-0.11.2/scripts/SneakerNet.roRun.pl line 295.
Use of uninitialized value $R1 in concatenation (.) or string at /SneakerNet-0.11.2/scripts/SneakerNet.roRun.pl line 296.
Use of uninitialized value $R2 in concatenation (.) or string at /SneakerNet-0.11.2/scripts/SneakerNet.roRun.pl line 296.
ERROR: R1 and R2 are not both represented as  and  at /SneakerNet-0.11.2/scripts/SneakerNet.roRun.pl line 296.
```
