### 2020-09-17 #2334 @lskatz
Hi @erinyoung ,
@kapsakcj and I are trying to replicate your error.  Could you also post your `singularity shell` command?  Thank you!  One concern we might have is that singularity might be enforcing a read only hard drive, but it is only a guess right now.
