### 2020-09-18 9e8d538
avoiding bug where it would crash if no fastq files were found for an entry
