### 2020-09-18 3ae82b3
fixed tempdir bug in SneakerNetPlugins
