### 2020-09-18 e55624e
Dockerfile version update
