### 2020-09-18 7609f8a
Trying a multistage build for docker
