+++ Dockerfile
@@ -1,18 +1,99 @@
-FROM ubuntu:xenial
+FROM ubuntu:bionic
 
-LABEL base.image="ubuntu:xenial"
-LABEL container.version="1"
+LABEL base.image="ubuntu:bionic"
+LABEL dockerfile.version="1.0.0"
 LABEL software="SneakerNet"
-LABEL software.version="0.11.6"
-LABEL description="QA/QC pipeline for a MiSeq/HiSeq/Ion Torrent run"
+LABEL software.version="0.11.5"
+LABEL description="SneakerNet QA/QC system for primary genomic data"
 LABEL website="https://github.com/lskatz/SneakerNet"
-LABEL license="https://github.com/lskatz/SneakerNet/blob/master/LICENSE"
-LABEL maintainer="Curtis Kapsak"
-LABEL maintainer.email="pjx8@cdc.gov"
+LABEL license="https://github.com/lskatz/SneakerNet/LICENSE.md"
+LABEL maintainer1="Lee Katz"
+LABEL maintainer1.email="gzu2@cdc.gov"
+LABEL maintainer2="Curtis Kapsak"
+LABEL maintainer2.email="pjx8@cdc.gov"
 
-### install dependencies ###
-# libexpat1-dev needed for cpanm to install XML::Parser and other perl modules
-RUN apt-get update && apt-get -y --no-install-recommends install \
+# https://github.com/StaPH-B/docker-builds/blob/master/mash/2.2/Dockerfile
+FROM staphb/mash:2.2 AS mash
+# https://github.com/StaPH-B/docker-builds/blob/master/skesa/2.4.0/Dockerfile
+FROM staphb/skesa:2.4.0 AS skesa
+# etc on the URL sources
+FROM staphb/kraken:1.1.1-no-db AS kraken
+FROM staphb/mlst:2.19.0 AS mlst
+FROM staphb/prokka:1.14.5 AS prokka
+FROM staphb/shovill:1.1.0 AS shovill
+FROM staphb/seqtk:1.3 AS seqtk
+FROM staphb/staramr:0.7.1 AS staramr
+FROM staphb/salmid:0.1.23 AS salmid
+#FROM rust:1.46.0-slim-buster AS rust
+
+# Other sources
+FROM mgibio/samtools:1.9 AS samtools
+FROM flowcraft/krona:2.7-1 AS krona
+FROM ummidock/chewbbaca:2.1.0-1 AS chewbbaca
+
+# EDIT: this bioperl container uses perl/5.18 which doesn't match our perl v5.26.1
+# Bring in libraries
+#FROM bioperl/bioperl:release-1-7-2 AS bioperl
+
+# "Import" ubuntu:bionic one more time as a hack.
+# I hope this step kind of clears the slate so that the COPY
+# commands don't fubar.
+FROM ubuntu:bionic
+
+# Let's get all the executables into our path
+COPY --from=mash      /mash-Linux64-v2.2/* /usr/local/bin/
+COPY --from=skesa     /skesa/*             /usr/local/bin/
+COPY --from=kraken    /opt/kraken/*        /usr/local/bin/
+COPY --from=kraken    /opt/bin/*           /usr/local/bin/
+COPY --from=mlst      /mlst-2.19.0         /mlst-2.19.0    
+COPY --from=mlst      /usr/local/bin/any2fasta              /usr/local/bin/
+COPY --from=prokka    /bedtools2/bin       /usr/local/bin/
+COPY --from=prokka    /prokka-1.14.5/bin   /usr/local/bin/
[…]
