### 2020-09-18 e919b74
Trying a multistage build for docker
