### 2020-09-18 80f262f
Figuring out docker push via travis
