+++ .travis.yml
@@ -91,3 +91,16 @@ script:
   - cpanm --installdeps --notest .
   - perl Makefile.PL
   - make test
+
+# deploying to dockerhub
+deploy:
+  on:
+    branch: master
+  provider: script
+  script:
+    - docker build -t $CI_PROJECT_NAME:$TRAVIS_REPO_SLUG .
+    - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+    - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
+    - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
+    - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+    - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:latest 
