+++ .travis.yml
@@ -2,7 +2,8 @@ language: perl
 os: linux
 perl:
   - "5.24-shrplib"
-script: 
+script: true
+install:
   - echo "Just testing docker on this branch"
 
 # deploying to dockerhub
