+++ .travis.yml
@@ -13,9 +13,9 @@ deploy:
   provider: script
   script:
     - CI_PROJECT_NAME=SneakerNet
-    - [[ -z "$DOCKER_USERNAME" ]] && echo "DOCKER_USERNAME not set"
-    - [[ -z "$DOCKER_PASSWORD" ]] && echo "DOCKER_PASSWORD not set"
-    - [[ -z "$CI_PROJECT_NAME" ]] && echo "CI_PROJECT_NAME not set"
+#   - [[ -z "$DOCKER_USERNAME" ]] && echo "DOCKER_USERNAME not set"
+#   - [[ -z "$DOCKER_PASSWORD" ]] && echo "DOCKER_PASSWORD not set"
+#   - [[ -z "$CI_PROJECT_NAME" ]] && echo "CI_PROJECT_NAME not set"
     - docker build -t $CI_PROJECT_NAME:$TRAVIS_REPO_SLUG .
     - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
     - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
