### 2020-09-19 df14da6
double checking that travis variables are set up
