### 2020-09-19 85637f3
double checking that travis variables are set up
