### 2020-09-19 97a2f8b
double checking that travis variables are set up
