+++ .travis.yml
@@ -8,20 +8,8 @@ install:
 
 # deploying to dockerhub
 deploy:
+  provider: script
+  script: bash .travis-ci/docker_push
   on:
     branch: $TRAVIS_BRANCH
-  provider: script
-  script:
-#   - [[ -z "$DOCKER_USERNAME" ]] && echo "DOCKER_USERNAME not set"
-#   - [[ -z "$DOCKER_PASSWORD" ]] && echo "DOCKER_PASSWORD not set"
-#   - [[ -z "$CI_PROJECT_NAME" ]] && echo "CI_PROJECT_NAME not set"
-    - CI_PROJECT_NAME=SneakerNet
-    - echo "$CI_PROJECT_NAME"
-    - echo ${DOCKER_USERNAME:0:2}
-    - echo ${DOCKER_PASSWORD:0:2}
-    - docker build -t $CI_PROJECT_NAME:$TRAVIS_REPO_SLUG .
-    - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
-    - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
-    - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
-    - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
-    - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:latest 
+
