+++ .travis.yml
@@ -6,10 +6,21 @@ script: true
 install:
   - echo "Just testing docker on this branch"
 
+after_success:
+  - CI_PROJECT_NAME=sneakernet
+  - echo "$CI_PROJECT_NAME"
+  - docker build -t $CI_PROJECT_NAME:$TRAVIS_REPO_SLUG .
+  - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+  - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
+  - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
+  - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+  - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:latest 
+
+
 # deploying to dockerhub
-deploy:
-  provider: script
-  script: bash .travis-ci/docker_push
-  on:
-    branch: $TRAVIS_BRANCH
+#deploy:
+#  provider: script
+#  script: bash .travis-ci/docker_push
+#  on:
+#    branch: $TRAVIS_BRANCH
 
