+++ .travis.yml
@@ -8,10 +8,11 @@ install:
 
 after_success:
   - CI_PROJECT_NAME=sneakernet
+  - SLUG=$(echo "$TRAVIS_REPO_SLUG" | tr '[:upper:]' '[:lower:]')
   - echo "$CI_PROJECT_NAME"
-  - docker build -t $CI_PROJECT_NAME:$TRAVIS_REPO_SLUG .
-  - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
-  - docker tag ${CI_PROJECT_NAME}:${TRAVIS_REPO_SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
+  - docker build -t $CI_PROJECT_NAME:$SLUG .
+  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
   - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
   - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
   - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:latest 
