+++ .travis.yml
@@ -10,7 +10,7 @@ after_success:
   - CI_PROJECT_NAME=sneakernet
   - SLUG=$(echo "$TRAVIS_REPO_SLUG" | tr '[:upper:]' '[:lower:]')
   - echo "$CI_PROJECT_NAME"
-  - docker build -t $CI_PROJECT_NAME:$SLUG .
+  - docker build -t $CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
   - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
   - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
   - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
