+++ .travis.yml
@@ -10,11 +10,11 @@ after_success:
   - CI_PROJECT_NAME=sneakernet
   - SLUG=$(echo "$TRAVIS_REPO_SLUG" | tr '[:upper:]' '[:lower:]')
   - echo "$CI_PROJECT_NAME"
-  - docker build -t $CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
-  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+  - docker build -t $CI_PROJECT_NAME:$TRAVIS_COMMIT .
+  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
   - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
-  - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_PULL_REQUEST_SHA
+  - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:latest 
 
 
