+++ .travis.yml
@@ -13,8 +13,9 @@ after_success:
   - echo "$CI_PROJECT_NAME"
   - echo $SLUG
   - echo $REPOSITORY_URL
-  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_COMMIT
-  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
+  - echo $TRAVIS_COMMIT
+  - docker tag ${CI_PROJECT_NAME}:${SLUG} $CI_PROJECT_NAME:$TRAVIS_COMMIT
+    #  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
   - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
   - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker push $REPOSITORY_URL/$CI_PROJECT_NAME:latest 
