+++ .travis.yml
@@ -14,6 +14,7 @@ after_success:
   - echo $SLUG
   - echo $REPOSITORY_URL
   - echo $TRAVIS_COMMIT
+  - echo docker tag ${CI_PROJECT_NAME}:${SLUG} $CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker tag ${CI_PROJECT_NAME}:${SLUG} $CI_PROJECT_NAME:$TRAVIS_COMMIT
     #  - docker tag ${CI_PROJECT_NAME}:${SLUG} $REPOSITORY_URL/$CI_PROJECT_NAME:latest
   - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
