+++ .travis.yml
@@ -14,9 +14,9 @@ after_success:
   - echo $SLUG
   - echo $REPOSITORY_URL
   - echo $TRAVIS_COMMIT
-  - echo docker tag ${SLUG}:${CI_PROJECT_NAME} $CI_PROJECT_NAME:$TRAVIS_COMMIT
-  -      docker tag ${SLUG}:${CI_PROJECT_NAME} $CI_PROJECT_NAME:$TRAVIS_COMMIT
-  -      docker tag ${SLUG}:${CI_PROJECT_NAME} $CI_PROJECT_NAME:latest
+  - echo docker tag $CI_PROJECT_NAME:$TRAVIS_COMMIT $CI_PROJECT_NAME:$TRAVIS_COMMIT
+  -      docker tag $CI_PROJECT_NAME:$TRAVIS_COMMIT $CI_PROJECT_NAME:$TRAVIS_COMMIT
+  -      docker tag $CI_PROJECT_NAME:$TRAVIS_COMMIT $CI_PROJECT_NAME:latest
   - echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin 
   - docker push $CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker push $CI_PROJECT_NAME:latest 
