### 2020-09-19 a53678b
fixing docker login
