### 2020-09-19 534e731
trying to get secret env variables for docker login
