+++ .travis.yml
@@ -14,14 +14,18 @@ after_success:
   - echo $SLUG
   - echo $REPOSITORY_URL
   - echo $TRAVIS_COMMIT
+  - DOCKER_PASSWORD=${{secrets.DOCKER_PASSWORD}}
+  - DOCKER_USERNAME=${{secrets.DOCKER_USERNAME}}
+  - echo $DOCKER_USERNAME
   - echo docker tag $CI_PROJECT_NAME:$TRAVIS_COMMIT $SLUG:$TRAVIS_COMMIT
-# e.g., docker tag httpd:test fedora/httpd:version1.0.test
   -      docker tag $CI_PROJECT_NAME:$TRAVIS_COMMIT $SLUG:$TRAVIS_COMMIT
   -      docker tag $CI_PROJECT_NAME:$TRAVIS_COMMIT $SLUG:latest
-  - echo ${{secrets.DOCKER_PASSWORD}} | docker login --username "${{secrets.DOCKER_USERNAME}}" --password-stdin 
+  - echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin 
   - docker push $CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker push $CI_PROJECT_NAME:latest 
 
+cache:
+  - 
 
 # deploying to dockerhub
 #deploy:
