+++ .travis.yml
@@ -24,8 +24,14 @@ after_success:
   - docker push $CI_PROJECT_NAME:$TRAVIS_COMMIT
   - docker push $CI_PROJECT_NAME:latest 
 
+https://stackoverflow.com/a/58505735
 cache:
-  - 
+  directories:
+    - docker_images
+before_install:
+  - docker load -i docker_images/images.tar || true
+before_cache:
+  -docker save -o docker_images/images.tar $(docker images -a -q)
 
 # deploying to dockerhub
 #deploy:
