+++ .travis.yml
@@ -31,7 +31,7 @@ cache:
 before_install:
   - docker load -i docker_images/images.tar || true
 before_cache:
-  -docker save -o docker_images/images.tar $(docker images -a -q)
+  - docker save -o docker_images/images.tar $(docker images -a -q)
 
 # deploying to dockerhub
 #deploy:
