+++ .travis.yml
@@ -25,13 +25,12 @@ after_success:
   - docker push $CI_PROJECT_NAME:latest 
 
 #https://stackoverflow.com/a/58505735
-cache:
-  directories:
-    - docker_images
 before_install:
-  - docker load -i docker_images/images.tar || true
+  - if [[ -d $HOME/docker ]]; then \ls $HOME/docker/*.tar.gz | xargs -I {file} sh -c "zcat {file} | docker load"; fi
 before_cache:
-  - docker save -o docker_images/images.tar $(docker images -a -q)
+  - mkdir -p $HOME/docker
+  - docker images -a --filter='dangling=false' --format '{{.Repository}}:{{.Tag}} {{.ID}} > $HOME/dockerImages.stdout
+  - xargs -n 2 -t sh -c 'test -e $HOME/docker/$1.tar.gz || docker save $0 | gzip -2 > $HOME/docker/$1.tar.gz' < $HOME/dockerImages.stdout
 
 # deploying to dockerhub
 #deploy:
