### 2020-09-19 2eb800e
view dockerImages.stdout
