+++ .travis.yml
@@ -31,8 +31,13 @@ before_cache:
   - mkdir -p $HOME/docker
   - docker images -a --filter='dangling=false' --format '{{.Repository}}:{{.Tag}} {{.ID}}' > $HOME/dockerImages.stdout
   - cat $HOME/dockerImages.stdout
+  - echo "0 0" >> $HOME/dockerImages.stdout # avoid not finding 'sh'
   - xargs -n 2 -t sh -c 'test -e $HOME/docker/$1.tar.gz || docker save $0 | gzip -2 > $HOME/docker/$1.tar.gz' < $HOME/dockerImages.stdout
-
+cache:
+  bundler: true
+  directories:
+    - $HOME/docker
+    
 # deploying to dockerhub
 #deploy:
 #  provider: script
