### 2020-09-19 480463f
test for size of dockerImages.stdout
