+++ .travis.yml
@@ -29,10 +29,10 @@ before_install:
   - if [[ -d $HOME/docker ]]; then \ls $HOME/docker/*.tar.gz | xargs -I {file} sh -c "zcat {file} | docker load"; fi
 before_cache:
   - mkdir -p $HOME/docker
+  - rm -vf $HOME/dockerImages.stdout
   - docker images -a --filter='dangling=false' --format '{{.Repository}}:{{.Tag}} {{.ID}}' > $HOME/dockerImages.stdout
   - cat $HOME/dockerImages.stdout
-  - echo "0 0" >> $HOME/dockerImages.stdout # avoid not finding 'sh'
-  - xargs -n 2 -t sh -c 'test -e $HOME/docker/$1.tar.gz || docker save $0 | gzip -2 > $HOME/docker/$1.tar.gz' < $HOME/dockerImages.stdout
+  - if [[ -s $HOME/dockerImages.stdout ]]; then xargs -n 2 -t sh -c 'test -e $HOME/docker/$1.tar.gz || docker save $0 | gzip -2 > $HOME/docker/$1.tar.gz' < $HOME/dockerImages.stdout; fi
 cache:
   bundler: true
   directories:
