### 2020-09-19 cd447ec
from xargs to bash loop
