+++ .travis.yml
@@ -26,7 +26,7 @@ after_success:
 
 #https://stackoverflow.com/a/58505735
 before_install:
-  - if [[ -d $HOME/docker ]]; then \ls $HOME/docker/*.tar.gz | xargs -I {file} sh -c "zcat {file} | docker load"; fi
+  - for file in $HOME/docker/*.tar.gz; do echo "loading $file"; zcat "$file" | docker load"; done
 before_cache:
   - mkdir -p $HOME/docker
   - rm -vf $HOME/dockerImages.stdout
