+++ .travis.yml
@@ -26,6 +26,7 @@ after_success:
 
 #https://stackoverflow.com/a/58505735
 before_install:
+  - rm -vf $HOME/dockerImages.stdout
   - for file in $HOME/docker/*.tar.gz; do echo "loading $file"; zcat "$file" | docker load; done
 before_cache:
   - mkdir -p $HOME/docker
