### 2020-09-19 22e1c20
checking if file exists
