### 2020-09-21 a6e95f5
updating run commands for variable printing
