### 2020-09-21 a41a613
updating run commands for variable printing
