### 2020-09-21 a6fc922
updating run commands for variable printing
