+++ .github/workflows/travis-docker.yml
@@ -21,7 +21,7 @@ jobs:
     steps:
       - id:  test-not-so-secret-variable
         run: echo "logging in with ${DOCKER_USERNAME:0:6}"
-      - id:  test docker tag
+      - id:  test-docker-tag
         run: echo "$CI_PROJECT_NAME:$GITHUB_SHA"
       #- id:  docker-login-ssh-key
       #  run: docker login --username "$DOCKER_USERNAME"
