### 2020-09-21 bee3267
more debugging; switching build directory to HOME
