+++ .github/workflows/travis-docker.yml
@@ -23,9 +23,13 @@ jobs:
         run: echo "logging in with ${DOCKER_USERNAME:0:6}"
       - id:  test-docker-tag
         run: echo "$CI_PROJECT_NAME:$GITHUB_SHA"
-      #- id:  docker-login-ssh-key
-      #  run: docker login --username "$DOCKER_USERNAME"
+      - id:  test-other-vars
+        run: echo "HOME:$HOME PWD:$PWD"
+      - id:  test-other-env
+        run: ls
       - id:  docker-login
         run: echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin; echo $?
       - id:  docker-build
         run: docker build -t $CI_PROJECT_NAME:$GITHUB_SHA .
+        working-directory: $HOME
+
