+++ .github/workflows/travis-docker.yml
@@ -26,8 +26,7 @@ jobs:
       - id:  test-other-vars
         run: echo "HOME:$HOME PWD:$PWD"
       - id:  test-other-env
-        run: ls -R
-        working-directory: $HOME
+        run: ls -R $HOME
       - id:  docker-login
         run: echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin; echo $?
       - id:  docker-build
