+++ .github/workflows/travis-docker.yml
@@ -15,9 +15,9 @@ jobs:
     runs-on: ubuntu-18.04
     steps:
       -uses: actions/checkout@v2
-      with:
-        fetch-depth: 1
         ref: docker-multistage
+        with:
+          fetch-depth: 1
       - id:  test-not-so-secret-variable
         run: echo "logging in with ${DOCKER_USERNAME:0:6}"
       - id:  test-docker-tag
