+++ .github/workflows/travis-docker.yml
@@ -15,8 +15,8 @@ jobs:
     runs-on: ubuntu-18.04
     steps:
       -uses: actions/checkout@v2
-        ref: docker-multistage
         with:
+          ref: docker-multistage
           fetch-depth: 1
       - id:  test-not-so-secret-variable
         run: echo "logging in with ${DOCKER_USERNAME:0:6}"
