### 2020-09-21 7383553
fixed working directory during docker build
