### 2020-09-21 c3263a2
fixed working directory during docker build
