+++ .github/workflows/travis-docker.yml
@@ -29,6 +29,6 @@ jobs:
       - id:  docker-login
         run: echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin; echo $?
       - id:  docker-build
-        run: ls -lhdA *; docker build -t $CI_PROJECT_NAME:$GITHUB_SHA .
+        run: cd $GITHUB_WORKSPACE && ls -lhA && docker build -t $CI_PROJECT_NAME:$GITHUB_SHA .
         working-directory: $GITHUB_WORKSPACE
 
