+++ .github/workflows/travis-docker.yml
@@ -23,14 +23,17 @@ jobs:
       - id:  test-docker-tag
         run: echo "$CI_PROJECT_NAME:$GITHUB_SHA"
       - id:  test-other-vars
-        run: echo "HOME:$HOME PWD:$PWD GITHUB_WORKSPACE:$GITHUB_WORKSPACE"
+        run: echo "HOME:$HOME PWD:$PWD GITHUB_WORKSPACE:$GITHUB_WORKSPACE $GITHUB_ACTOR/$CI_PROJECT_NAME:$GITHUB_SHA"
       - id:  test-github-root
         run: ls -R $GITHUB_WORKSPACE
-      - id:  docker-login
-        run: echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin; echo $?
       - id:  see-if-git-dir-is-good
         run: ls -lha $GITHUB_WORKSPACE
       - id:  docker-build
         run: docker build -t $CI_PROJECT_NAME:$GITHUB_SHA $GITHUB_WORKSPACE
-        #working-directory: $GITHUB_WORKSPACE
+      - id:  docker-tag-latest
+        run:  docker tag $CI_PROJECT_NAME:$GITHUB_SHA $GITHUB_ACTOR/$CI_PROJECT_NAME:$GITHUB_SHA
+      - id:  docker-login
+        run: echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin
+      - id:  docker-push-latest
+        run: docker push $CI_PROJECT_NAME:$GITHUB_SHA
 
