### 2020-09-22 ef8f184
moving env out of job and to global
