+++ .github/workflows/travis-docker.yml
@@ -54,8 +54,8 @@ jobs:
     runs-on: ubuntu-18.04
     needs: docker_login
     steps:
-      -id:  todo
-       run: echo "TODO"
+      - id:  todo
+        run: echo "TODO"
 
   docker_push:
     runs-on: ubuntu-18.04
