### 2020-09-22 3fb01ee
added some debugging
