### 2020-09-22 6bf8409
added init and debugging jobs
