### 2020-09-22 81d3d3b
figuring out variable continuity
