### 2020-09-22 96ceae5
figuring out variable continuity
