+++ .github/workflows/travis-docker.yml
@@ -21,6 +21,7 @@ jobs:
       - id:  set-remote-tag
         run: |
           REMOTE_TAG=$(echo "$GITHUB_REPOSITORY:$GITHUB_SHA" | tr '[:upper:]' '[:lower:]')
+          echo "$REMOTE_TAG"
           echo "::set-output name=key::$REMOTE_TAG"
 
   debugging:
