+++ .github/workflows/travis-docker.yml
@@ -40,9 +40,7 @@ jobs:
           echo "GITHUB_WORKSPACE:$GITHUB_WORKSPACE"
           echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $GITHUB_ACTOR/$CI_PROJECT_NAME:$GITHUB_SHA"
           echo "REMOTE_TAG $REMOTE_TAG"
-          echo ${{needs.init.outputs.set-remote-tag}}
-          echo ${{needs.init.outputs.key}}
-          echo ${{needs.init.outputs}}
+          echo ${{ toJson(needs.init.outputs) }}
       - id:  ls-all-the-things
         run: |
           ls -R $GITHUB_WORKSPACE
