### 2020-09-22 32ba999
debugging more with JSON
