### 2020-09-22 ba2a13d
moved git checkout to init job
