### 2020-09-22 71b17b1
maybe I needed to do fromJSON. And moved git checkout back.
