+++ .github/workflows/travis-docker.yml
@@ -38,8 +38,7 @@ jobs:
           echo "GITHUB_WORKSPACE:$GITHUB_WORKSPACE"
           echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $GITHUB_ACTOR/$CI_PROJECT_NAME:$GITHUB_SHA"
           echo "REMOTE_TAG $REMOTE_TAG"
-          echo ${{ toJson(needs.init.outputs) }}
-          echo ${{ fromJson(needs.init.outputs) }}
+          echo ${{ toJson(needs.init) }}
 
   docker_build:
     runs-on: ubuntu-18.04
