+++ .github/workflows/travis-docker.yml
@@ -16,12 +16,12 @@ jobs:
   init:
     runs-on: ubuntu-18.04
     outputs:
-      remote-tag: ${{ steps.set-remote-tag.outputs.key }}
+      remote-tag: ${{ steps.set-remote-tag.outputs.remote_tag}}
     steps:
       - id:  lowercase-slug
         run: SLUG=$(echo $GITHUB_REPOSITORY:$GITHUB_SHA | tr '[:upper:]' '[:lower:]')
       - id:  set-remote-tag
-        run: echo "::set-output name=REMOTE_TAG::$SLUG"
+        run: echo "::set-output name=remote_tag::$SLUG"
 
   debugging:
     runs-on: ubuntu-18.04
@@ -38,7 +38,7 @@ jobs:
           echo "GITHUB_WORKSPACE:$GITHUB_WORKSPACE"
           echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $GITHUB_ACTOR/$CI_PROJECT_NAME:$GITHUB_SHA"
           echo "REMOTE_TAG $REMOTE_TAG"
-          echo ${{ toJson(needs.init) }} 
+          echo ${{ toJson(needs.init.outputs) }} 
 
   docker_build:
     runs-on: ubuntu-18.04
