+++ .github/workflows/travis-docker.yml
@@ -27,7 +27,7 @@ jobs:
     runs-on: ubuntu-18.04
     needs:   init
     env:     
-      REMOTE_TAG: ${{needs.init.outputs.set-remote-tag}}
+      REMOTE_TAG: ${{needs.init.outputs.remote-tag}}
     steps:
       - id:  print-globals
         run: |
