+++ .github/workflows/travis-docker.yml
@@ -20,6 +20,8 @@ jobs:
     steps:
       - id:  lowercase-slug
         run: SLUG=$(echo $GITHUB_REPOSITORY:$GITHUB_SHA | tr '[:upper:]' '[:lower:]')
+      - id:  test-slug
+        run: echo $SLUG
       - id:  set-remote-tag
         run: echo "::set-output name=remote_tag::$SLUG"
 
