+++ .github/workflows/travis-docker.yml
@@ -7,32 +7,24 @@ on:
       - docker-multistage
 
 env:      
-  DOCKER_PASSWORD: ${{  secrets.DOCKER_PASSWORD }}
-  DOCKER_USERNAME: lskatz
-  #DOCKER_USERNAME: ${{  secrets.DOCKER_USERNAME }}
-  CI_PROJECT_NAME: sneakernet
-  REMOTE_TAG:      UNDEFINED
+  DOCKER_PASSWORD:   ${{  secrets.DOCKER_PASSWORD }}
+  DOCKER_USERNAME:   lskatz
+  #DOCKER_USERNAME:  ${{  secrets.DOCKER_USERNAME }}
+  CI_PROJECT_NAME:   sneakernet
   github_repository: lskatz/sneakernet
-  SLUG:            "lskatz/sneakernet:$GITHUB_SHA"
+  SLUG:              format('lskatz/sneakernet:{0}', $GITHUB_SHA)
+  LOCAL_TAG_LATEST:  format('sneakernet:{0}', $GITHUB_SHA)
 
 jobs:
   init:
     runs-on: ubuntu-18.04
-    outputs:
-      remotetag: ${{ steps.setremotetag.outputs.remotetag}}
     steps:
-      - id:  lowercase-slug
-        run: export SLUG=$(echo $GITHUB_REPOSITORY:$GITHUB_SHA | tr '[:upper:]' '[:lower:]'); echo $SLUG
-      - id:  test-slug
-        run: echo $SLUG
-      - id:  setremotetag
-        run: echo "::set-output name=remotetag::$SLUG"
+      - id:  todo
+        run: echo "todo"
 
   debugging:
     runs-on: ubuntu-18.04
     needs:   init
-    env:     
-      REMOTE_TAG: ${{needs.init.outputs.remotetag}}
     steps:
       - id:  print-globals
         run: |
@@ -41,9 +33,7 @@ jobs:
           echo "HOME:$HOME"
           echo "PWD:$PWD"
           echo "GITHUB_WORKSPACE:$GITHUB_WORKSPACE"
-          echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $GITHUB_ACTOR/$CI_PROJECT_NAME:$GITHUB_SHA"
-          echo "REMOTE_TAG $REMOTE_TAG"
-          echo ${{ toJson(needs.init.outputs.remotetag) }} 
+          echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $SLUG"
 
   docker_build:
     runs-on: ubuntu-18.04
@@ -59,39 +49,9 @@ jobs:
           ls -lha $GITHUB_WORKSPACE
       - id:  docker-build
         run: docker build -t $CI_PROJECT_NAME:$GITHUB_SHA $GITHUB_WORKSPACE
-
-  docker_login:
-    runs-on: ubuntu-18.04
-    needs: docker_build
-    steps:
-      - id:  docker-login
-        run: echo "$DOCKER_PASSWORD" | docker login --username "$DOCKER_USERNAME" --password-stdin
-
-  docker_build_latest:
-    runs-on: ubuntu-18.04
-    needs: docker_login
-    steps:
-      - id:  debug-env
-        run: |
-          echo $GITHUB_WORKSPACE
-          ls $GITHUB_WORKSPACE
       - id:  docker-tag-latest
         run: |
-          REMOTE_TAG=$(echo "$GITHUB_REPOSITORY:$GITHUB_SHA" | tr '[:upper:]' '[:lower:]')
-          docker tag $CI_PROJECT_NAME:$GITHUB_SHA $REMOTE_TAG
-
[…]
