+++ .github/workflows/travis-docker.yml
@@ -13,7 +13,7 @@ env:
   CI_PROJECT_NAME:   sneakernet
   github_repository: lskatz/sneakernet
   SLUG:              ${{ format('lskatz/sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
-  LOCAL_TAG_LATEST:  ${{ format('sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
+  #LOCAL_TAG_LATEST:  ${{ format('sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
 
 jobs:
   init:
