+++ .github/workflows/travis-docker.yml
@@ -9,10 +9,10 @@ on:
 env:      
   DOCKER_PASSWORD:   ${{  secrets.DOCKER_PASSWORD }}
   DOCKER_USERNAME:   lskatz
-  #DOCKER_USERNAME:  ${{  secrets.DOCKER_USERNAME }}
+  DOCKER_USERNAME_SECRET:  ${{  secrets.DOCKER_USERNAME }}
   CI_PROJECT_NAME:   sneakernet
   github_repository: lskatz/sneakernet
-  SLUG:              ${{ format('lskatz/sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
+  SLUG:              ${{ format('lskatz/sneakernet:{0}', secrets.DOCKER_USERNAME ) }}
   #LOCAL_TAG_LATEST:  ${{ format('sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
 
 jobs:
