### 2020-09-22 9bcfcf5
just making the variables work
