+++ .github/workflows/travis-docker.yml
@@ -12,7 +12,6 @@ env:
   #DOCKER_USERNAME_SECRET:  ${{  secrets.DOCKER_USERNAME }}
   CI_PROJECT_NAME:   sneakernet
   github_repository: lskatz/sneakernet
-  GITHUB_SHA:        ${{ env.GITHUB_SHA }}
   #SLUG:              ${{ format('lskatz/sneakernet:{0}', GITHUB_SHA ) }}
   #LOCAL_TAG_LATEST:  ${{ format('sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
 
@@ -34,7 +33,7 @@ jobs:
           echo "HOME:$HOME"
           echo "PWD:$PWD"
           echo "GITHUB_WORKSPACE:$GITHUB_WORKSPACE"
-          echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $SLUG"
+          echo "GITHUB_ACTOR/CI_PROJECT_NAME:GITHUB_SHA $github_repository:$GITHUB_SHA"
 
   docker_build:
     runs-on: ubuntu-18.04
@@ -52,7 +51,7 @@ jobs:
         run: docker build -t $CI_PROJECT_NAME:$GITHUB_SHA $GITHUB_WORKSPACE
       - id:  docker-tag-latest
         run: |
-          docker tag $CI_PROJECT_NAME:$GITHUB_SHA $SLUG
+          docker tag $CI_PROJECT_NAME:$GITHUB_SHA $github_repository:$GITHUB_SHA
       - id:  docker-push-latest
         run: docker push $CI_PROJECT_NAME:$GITHUB_SHA
 
