+++ .github/workflows/travis-docker.yml
@@ -43,6 +43,8 @@ jobs:
         with:
           ref: docker-multistage
           fetch-depth: 1
+      - uses: satackey/action-docker-layer-caching@v0.0.8
+          continue-on-error: true
       - id:  ls-all-the-things
         run: |
           ls -R $GITHUB_WORKSPACE
