+++ .github/workflows/travis-docker.yml
@@ -54,7 +54,7 @@ jobs:
         run: docker build . --file $GITHUB_WORKSPACE/Dockerfile --tag $CI_PROJECT_NAME:$GITHUB_SHA
       - id:  docker-tag-latest
         run: |
-          docker tag $CI_PROJECT_NAME:$GITHUB_SHA $github_repository:$GITHUB_SHA .
+          docker tag $CI_PROJECT_NAME:$GITHUB_SHA $github_repository:$GITHUB_SHA
       - id:  docker-push-latest
         run: docker push $CI_PROJECT_NAME:$GITHUB_SHA
 
