### 2020-09-23 70100c5
added docker login
