+++ .github/workflows/github-docker.yml
@@ -20,7 +20,6 @@ jobs:
   push_to_registry:
     name: Push Docker image to Docker Hub
     runs-on: ubuntu-18.04
-    needs: init
     steps:
       - name: Check out the repo
         uses: actions/checkout@v2
