+++ .github/workflows/github-docker.yml
@@ -28,6 +28,6 @@ jobs:
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-          repository: my-docker-hub-namespace/my-docker-hub-repository
+          repository: lskatz/sneakernet
           tag_with_ref: true
 
