### 2020-09-25 91e7c14
pushing to both docker and github
https://docs.github.com/en/actions/guides/publishing-docker-images
