### 2020-09-25 e4c9acd
renamed the github container name from my-image to sneakernet
