+++ .github/workflows/github-docker.yml
@@ -29,7 +29,8 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: lskatz/sneakernet
-          tag_with_ref: true
+          tags: latest
+          #tag_with_ref: true
       - name: Push to GitHub Packages
         uses: docker/build-push-action@v1
         with:
@@ -37,4 +38,5 @@ jobs:
           password: ${{ secrets.GITHUB_TOKEN }}
           registry: docker.pkg.github.com
           repository: lskatz/sneakernet/sneakernet
-          tag_with_ref: true
+          tags: latest
+          #tag_with_ref: true
