+++ Dockerfile
@@ -60,12 +60,17 @@ COPY --from=chewbbaca /NGStools/clustalw-2.1-linux-x86_64-libcppstatic  /NGStool
 COPY --from=chewbbaca /NGStools/Prodigal                                /NGStools
 COPY --from=chewbbaca /NGStools/prodigal_training_files                 /NGStools
 COPY --from=chewbbaca /usr/local/bin/*     /usr/local/bin/
-COPY --from=blast     /blast                /blast    
+COPY --from=blast     /blast               /blast    
+
 #COPY --from=rust      /usr/local/rustup    /usr/local/rustup
 #COPY --from=rust      /usr/local/cargo     /usr/local/cargo
 
 # Libraries
 COPY --from=staramr   /usr/local/lib/python3.6             /usr/local/lib/python3.6/
+COPY --from=blast     /lib/x86_64-linux-gnu /lib/x86_64-linux-gnu
+COPY --from=blast     /lib64               /lib64
+COPY --from=blast     /usr/lib/x86_64-linux-gnu  /usr/lib/x86_64-linux-gnu
+
 # Taking a risk using python3.5 libraries in a python3.6 folder
 COPY --from=salmid    /usr/local/lib/python3.5             /usr/local/lib/python3.6/
 COPY --from=chewbbaca /usr/local/lib/python3.5             /usr/local/lib/python3.6/
