### 2020-10-02 a261a2a
getting blast from mlst container
