### 2020-10-02 7a6ffd3
manual installation of staramr
