### 2020-10-02 2549cce
manual installation of staramr2
