+++ Dockerfile
@@ -51,7 +51,7 @@ COPY --from=prokka    /bedtools2/bin       /usr/local/bin/
 COPY --from=prokka    /prokka-1.14.5/bin   /usr/local/bin/
 COPY --from=prokka    /barrnap-0.9/bin     /usr/local/bin/
 COPY --from=seqtk     /seqtk-1.3           /usr/local/bin/
-COPY --from=staramr   /usr/local/bin       /usr/local/bin/
+#COPY --from=staramr   /usr/local/bin       /usr/local/bin/
 COPY --from=salmid    /usr/local/bin       /usr/local/bin/
 COPY --from=samtools  /opt/samtools/bin/samtools            /usr/local/bin/
 COPY --from=krona     /NGStools/KronaTools-2.7              /NGStools/KronaTools-2.7
