+++ .github/workflows/github-docker.yml
@@ -31,7 +31,7 @@ jobs:
           repository: lskatz/sneakernet
           tags: |
             latest
-            ${{ GITHUB_SHA }}
+            ${{ github.GITHUB_SHA }}
           #tag_with_ref: true
       - name: Push to GitHub Packages
         uses: docker/build-push-action@v1
@@ -42,5 +42,5 @@ jobs:
           repository: lskatz/sneakernet/sneakernet
           tags: |
             latest
-            ${{ GITHUB_SHA }}
+            ${{ github.GITHUB_SHA }}
           #tag_with_ref: true
