### 2020-10-02 35ff62c
adding apt-get packages for staramr
