+++ .github/workflows/github-docker.yml
@@ -17,7 +17,7 @@ env:
   #LOCAL_TAG_LATEST:  ${{ format('sneakernet:{0}', ${{ env.GITHUB_SHA }} ) }}
 
 jobs:
-  push_to_registry:
+  push_to_registry_v1:
     name: Push Docker image to multiple registries
     runs-on: ubuntu-18.04
     steps:
@@ -35,6 +35,33 @@ jobs:
           #tag_with_ref: true
       - name: Push to GitHub Packages
         uses: docker/build-push-action@v1
+        with:
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+          registry: docker.pkg.github.com
+          repository: lskatz/sneakernet/sneakernet
+          tags: |
+            latest
+            ${{ github.GITHUB_SHA }}
+          #tag_with_ref: true
+  push_to_registry_v2:
+    name: Push Docker image to multiple registries
+    runs-on: ubuntu-18.04
+    steps:
+      - name: Check out the repo
+        uses: actions/checkout@v2
+      - name: Push to Docker Hub
+        uses: docker/build-push-action@v2
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+          repository: lskatz/sneakernet
+          tags: |
+            latest
+            ${{ github.GITHUB_SHA }}
+          #tag_with_ref: true
+      - name: Push to GitHub Packages
+        uses: docker/build-push-action@v2
         with:
           username: ${{ github.actor }}
           password: ${{ secrets.GITHUB_TOKEN }}
