+++ .github/workflows/github-docker.yml
@@ -18,7 +18,7 @@ env:
 
 jobs:
   push_to_registry_v1:
-    name: Push Docker image to multiple registries
+    name: Push Docker image to multiple registries v1
     runs-on: ubuntu-18.04
     steps:
       - name: Check out the repo
@@ -45,29 +45,29 @@ jobs:
             ${{ github.GITHUB_SHA }}
           #tag_with_ref: true
   push_to_registry_v2:
-    name: Push Docker image to multiple registries
+    name: Push Docker image to multiple registries v2
     runs-on: ubuntu-18.04
     steps:
       - name: Check out the repo
         uses: actions/checkout@v2
-      - name: Push to Docker Hub
-        uses: docker/build-push-action@v2
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to DockerHub
+        uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
-          repository: lskatz/sneakernet
-          tags: |
-            latest
-            ${{ github.GITHUB_SHA }}
-          #tag_with_ref: true
-      - name: Push to GitHub Packages
+      - name: Build and push
         uses: docker/build-push-action@v2
         with:
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-          registry: docker.pkg.github.com
-          repository: lskatz/sneakernet/sneakernet
-          tags: |
+          context: .
+          file: ./Dockerfile
+          pull: true
+          push: true
+          #build-args:
+          cache-from: type=registry,ref=lskatz/sneakernet
+          cache-to:   type=registry,ref=lskatz/sneakernet
+          tags:       |
             latest
-            ${{ github.GITHUB_SHA }}
-          #tag_with_ref: true
+            #lskatz/sneakernet:${{ github.GITHUB_SHA }}
+            #
