### 2020-10-05 4d0ca65
numpy installation through pip now
