### 2020-10-05 f541da1
Removing multistage for pip-based images
