+++ Dockerfile
@@ -29,7 +29,7 @@ FROM staphb/salmid:0.1.23 AS salmid
 # Other sources
 FROM mgibio/samtools:1.9 AS samtools
 FROM flowcraft/krona:2.7-1 AS krona
-FROM mickaelsilva/chewbbaca_py3 AS chewbbaca
+#FROM mickaelsilva/chewbbaca_py3 AS chewbbaca
 
 # EDIT: this bioperl container uses perl/5.18 which doesn't match our perl v5.26.1
 # Bring in libraries
@@ -52,13 +52,13 @@ COPY --from=prokka    /prokka-1.14.5/bin   /usr/local/bin/
 COPY --from=prokka    /barrnap-0.9/bin     /usr/local/bin/
 COPY --from=seqtk     /seqtk-1.3           /usr/local/bin/
 #COPY --from=staramr   /usr/local/bin       /usr/local/bin/
-COPY --from=salmid    /usr/local/bin       /usr/local/bin/
+#COPY --from=salmid    /usr/local/bin       /usr/local/bin/
 COPY --from=samtools  /opt/samtools/bin/samtools            /usr/local/bin/
 COPY --from=krona     /NGStools/KronaTools-2.7              /NGStools/KronaTools-2.7
-COPY --from=chewbbaca /NGStools/clustalw-2.1-linux-x86_64-libcppstatic  /NGStools
-COPY --from=chewbbaca /NGStools/Prodigal                                /NGStools
-COPY --from=chewbbaca /NGStools/prodigal_training_files                 /NGStools
-COPY --from=chewbbaca /usr/local/bin/*     /usr/local/bin/
+#COPY --from=chewbbaca /NGStools/clustalw-2.1-linux-x86_64-libcppstatic  /NGStools
+#COPY --from=chewbbaca /NGStools/Prodigal                                /NGStools
+#COPY --from=chewbbaca /NGStools/prodigal_training_files                 /NGStools
+#COPY --from=chewbbaca /usr/local/bin/*     /usr/local/bin/
 COPY --from=mlst      /ncbi-blast-2.9.0+   /ncbi-blast-2.9.0+/
 
 #COPY --from=rust      /usr/local/rustup    /usr/local/rustup
@@ -71,8 +71,8 @@ COPY --from=mlst      /ncbi-blast-2.9.0+   /ncbi-blast-2.9.0+/
 #COPY --from=blast     /usr/lib/x86_64-linux-gnu  /usr/lib/x86_64-linux-gnu
 
 # Taking a risk using python3.5 libraries in a python3.6 folder
-COPY --from=salmid    /usr/local/lib/python3.5             /usr/local/lib/python3.6/
-COPY --from=chewbbaca /usr/local/lib/python3.5             /usr/local/lib/python3.6/
+#COPY --from=salmid    /usr/local/lib/python3.5             /usr/local/lib/python3.6/
+#COPY --from=chewbbaca /usr/local/lib/python3.5             /usr/local/lib/python3.6/
 #COPY --from=bioperl   /usr/lib/            /usr/lib
 #COPY --from=bioperl   /usr/local/lib/      /usr/local/lib
 #COPY --from=bioperl   /usr/share           /usr/share
@@ -145,6 +145,9 @@ RUN apt-get update && \
  libcurl4-gnutls-dev \
  libssl-dev \
  libfindbin-libs-perl && \
+ psmisc && \
+ libatlas-base-dev && \
+ mafft && \
  apt-get autoclean && rm -rf /var/lib/apt/lists/* 
 #python-matplotlib ipython python-pandas python-sympy python-nose
 # python-numpy python-scipy  \
@@ -208,6 +211,9 @@ ENV PATH="${PATH}:\
[…]
