### 2020-10-05 10debae
fixed apt-get command syntax
