### 2020-10-05 #2334 @lskatz
We have one more thing that I hope will be an update.  I have been trying to make a multistage build for the container to make things easier.  It has been taking me a couple of weeks to learn this process and I'm not quite there yet... but I hope it will make the review process easier if it works 🤞.  If it works then hopefully it makes @erinyoung and @lfaller 's lives easier :)
