+++ .github/workflows/github-docker.yml
@@ -67,6 +67,7 @@ jobs:
           #build-args:
           cache-from: type=registry,ref=lskatz/sneakernet
           cache-to:   type=registry,ref=lskatz/sneakernet
-          tags:       lskatz/sneakernet:latest
-#lskatz/sneakernet:${{ github.GITHUB_SHA }}
+          tags:       |
+            lskatz/sneakernet:latest
+            lskatz/sneakernet:${{ github.GITHUB_SHA }}
 #
