+++ .github/workflows/github-docker.yml
@@ -69,5 +69,7 @@ jobs:
           cache-to:   type=registry,ref=lskatz/sneakernet
           tags:       |
             lskatz/sneakernet:latest
-            lskatz/sneakernet:${{ github.GITHUB_SHA }}
+            docker.pkg.github.com/lskatz/sneakernet:latest
+#           lskatz/sneakernet:${{ github.GITHUB_SHA }}
+#           docker.pkg.github.com/lskatz/sneakernet:${{ github.GITHUB_SHA }}
 #
