+++ Dockerfile
@@ -225,8 +225,8 @@ RUN staramr db update -d && staramr db info
 # Pip installations after I set the path
 # SalmID 0.1.23
 # apt deps: python-setuptools python3 python3-pip curl build-essential file git python3-venv
-RUN pip3 install poetry && \
- git clone https://github.com/hcdenbakker/SalmID.git --branch 0.1.23 --single-branch && \
+RUN pip3 install poetry
+RUN git clone https://github.com/hcdenbakker/SalmID.git --branch 0.1.23 --single-branch && \
  cd SalmID && \
  poetry build -vvv && \
  pip3 install dist/salmid*.whl
