### 2020-10-07 f072ecf
separate some pip3 installations for staramr; removed numpy pip3
