### 2020-10-09 e84cca7
removing staramr and salmid for right now
