### 2020-10-09 50ccfc5
removed all platforms except linux/amd64
