### 2020-10-09 5ebae80
uncomment pip commands
