### 2020-10-09 a57b108
combined some pip commands for fewer layers
