### 2020-10-12 4f13672
RUN echo "LC_ALL=C" >> /etc/environment
