### 2020-10-12 d1bc5bc
switched to UTF-8 per instructions
