### 2020-10-13 7f15f01
changed some cd to pushd
