### 2020-10-13 bb9c56e
Docker multistage (#48)
* testing docker from github

* minor fix to travis-ci

* changed branch to $TRAVIS_BRANCH

* adding env variables

* adding env variables

* double checking that travis variables are set up

* double checking that travis variables are set up

* double checking that travis variables are set up

* changing from deploy tag to after_success tag

* lowercase slug

* trying to fix the build tag name

* trying to fix the build tag name

* chewbbaca with python3

* chewbbaca with python3

* debugging docker tag

* debugging docker tag

* debugging docker tag

* debugging docke
[…]
