+++ Dockerfile
@@ -22,14 +22,14 @@ FROM staphb/mlst:2.19.0 AS mlst
 FROM staphb/prokka:1.14.5 AS prokka
 FROM staphb/shovill:1.1.0 AS shovill
 FROM staphb/seqtk:1.3 AS seqtk
-FROM staphb/staramr:0.7.1 AS staramr
+#FROM staphb/staramr:0.7.1 AS staramr
 FROM staphb/salmid:0.1.23 AS salmid
 #FROM rust:1.46.0-slim-buster AS rust
 
 # Other sources
 FROM mgibio/samtools:1.9 AS samtools
 FROM flowcraft/krona:2.7-1 AS krona
-FROM ummidock/chewbbaca:2.1.0-1 AS chewbbaca
+#FROM mickaelsilva/chewbbaca_py3 AS chewbbaca
 
 # EDIT: this bioperl container uses perl/5.18 which doesn't match our perl v5.26.1
 # Bring in libraries
@@ -51,19 +51,28 @@ COPY --from=prokka    /bedtools2/bin       /usr/local/bin/
 COPY --from=prokka    /prokka-1.14.5/bin   /usr/local/bin/
 COPY --from=prokka    /barrnap-0.9/bin     /usr/local/bin/
 COPY --from=seqtk     /seqtk-1.3           /usr/local/bin/
-COPY --from=staramr   /usr/local/bin       /usr/local/bin/
-COPY --from=salmid    /usr/local/bin       /usr/local/bin/
+#COPY --from=staramr   /usr/local/bin       /usr/local/bin/
+#COPY --from=salmid    /usr/local/bin       /usr/local/bin/
 COPY --from=samtools  /opt/samtools/bin/samtools            /usr/local/bin/
 COPY --from=krona     /NGStools/KronaTools-2.7              /NGStools/KronaTools-2.7
-COPY --from=chewbbaca /NGStools            /NGStools
-COPY --from=chewbbaca /usr/local/bin/*     /usr/local/bin/
+#COPY --from=chewbbaca /NGStools/clustalw-2.1-linux-x86_64-libcppstatic  /NGStools
+#COPY --from=chewbbaca /NGStools/Prodigal                                /NGStools
+#COPY --from=chewbbaca /NGStools/prodigal_training_files                 /NGStools
+#COPY --from=chewbbaca /usr/local/bin/*     /usr/local/bin/
+COPY --from=mlst      /ncbi-blast-2.9.0+   /ncbi-blast-2.9.0+/
+
 #COPY --from=rust      /usr/local/rustup    /usr/local/rustup
 #COPY --from=rust      /usr/local/cargo     /usr/local/cargo
 
 # Libraries
-COPY --from=staramr   /usr/local/lib/python3.6             /usr/local/lib/python3.6/
+#COPY --from=staramr   /usr/local/lib/python3.6             /usr/local/lib/python3.6/
+#COPY --from=blast     /lib/x86_64-linux-gnu /lib/x86_64-linux-gnu
+#COPY --from=blast     /lib64               /lib64
+#COPY --from=blast     /usr/lib/x86_64-linux-gnu  /usr/lib/x86_64-linux-gnu
+
 # Taking a risk using python3.5 libraries in a python3.6 folder
-COPY --from=salmid    /usr/local/lib/python3.5             /usr/local/lib/python3.6/
+#COPY --from=salmid    /usr/local/lib/python3.5             /usr/local/lib/python3.6/
+#COPY --from=chewbbaca /usr/local/lib/python3.5             /usr/local/lib/python3.6/
[…]
