+++ .github/workflows/build-and-push-container.yml
@@ -5,7 +5,7 @@ name: github-docker
 on: 
   push:
     branches:
-      - docker-multistage
+      - [ docker-multistage, master, main ]
 
 
 
