### 2020-10-13 #2334 @lskatz
Hi, I apologize for how long this took but we have a multistage Docker build.  Hopefully this makes the review much easier from here on out.  This container was tested only for the default workflow.  Please let us know if we need to do something more thorough.  Thanks to @kapsakcj once again for helping the containers efforts so much.

These changes are reflected in version 0.14.2

* https://github.com/lskatz/SneakerNet/releases/tag/v0.14.2
* https://github.com/lskatz/SneakerNet/runs/1248554962
* https://github.com/users/lskatz/packages/container/sneakernet/141925

@lpantano if the review process and journal allows it, I would like to add @kapsakcj as a co-first author.
