### 2020-10-22 #2334 @lfaller
Alright! I was able download the new version and run it and it did seem faster to me (though I didn't collect any hard numbers on that). FYI, I set the number of CPUs to 6, and that let me run the example in a reasonable amount of time.

Thanks for making these changes, I think it really improves usability!

My last nit is this:

I still find the documentation to be very confusing to read. If I was interested in finding a tool that could help me assemble/QC/analyze MiSeq data, this would not convince me that SneakerNet would be the way to go because it's not easy to get started quickly.

The main README.md currently looks like this:

![readme_screenshot](https://user-images.githubusercontent.com/7854149/96936296-a0b95c00-147a-11eb-8f67-f7e138cbf400.png)

1. The example code mentioned did not get me far -- I had to go dive deeply into the container documentation in order to get started and get things to run.

2. The synopsis text reads:
 > A pipeline for processing reads from a sequencing run.

 But from the main README, I don't know what tools are run in this pipeline. From reading more documentation, I understand
that the pipeline can be customized, but a description of the example pipeline would be nice because that is probably the
first starting point for a new user.

 The figure looks nice but it seems somewhat abstract. Is it supposed to describe the pipeline? If yes, I would expect arrows connecting the pieces. I assume the paper airplane means it's sending emails out? Or generates a report and sends it somewhere? I assume this because I've read the whole README multiple times but if I was new to this repo, it would seem unrelated.

3. The documentation is scattered. I agree that it's nice to have a README that's not 20 miles long, but none of the shorter sections in the README are self-explanatory, which means I have to click through to the "extended" documentation. I end up with a single tab open for every item in the `docs/` folder. Markdown is notoriously bad for organizing larger pieces of documentation, so a "read the docs" setup might be a nice addition.

Anyways, these are just some thoughts that popped into my head as I was trying to navigate the repo. Feel free to ignore 🙂
