### 2020-10-23 #2334 @lskatz
Thank you @lfaller !  We think that you had an excellent point that the documentation needs to be upfront on what it does specifically and so we added a [Main steps](https://github.com/lskatz/SneakerNet/blob/master/README.md#main-steps) section.

A large overhaul of the documentation might be difficult at this stage but hopefully that addresses at least, in a little part, your points 2 and 3.

Your point 1 is well taken that going through an example is really difficult to find in the documentation.  To that end, we have added a [quick start](https://github.com/lskatz/SneakerNet/blob/master/README.md#quick-start) and documentation on the [main script itself](https://github.com/lskatz/SneakerNet/blob/master/docs/SneakerNetPlugins.pl.md).  While not ideal, it glues together important pieces of documentation so that it is not scattered, as you correctly pointed out.
