### 2020-10-26 #2334 @lskatz
This new PDF just reflects the co-first author labels.  https://github.com/lskatz/SneakerNet/commit/0b0b84e182820dc20652a289a533e84c3a934009
