### 2020-10-29 e40a7dd
emoticons displayed
