+++ SneakerNet.plugins/sn_sarscov2_assembleAll.pl
@@ -10,15 +10,16 @@ use File::Temp qw/tempdir/;
 use File::Copy qw/mv cp/;
 use Bio::SeqIO;
 use Bio::FeatureIO::gff;
+use List::Util qw/min max/;
 
 use threads;
 use Thread::Queue;
 
 use FindBin qw/$RealBin/;
 use lib "$FindBin::RealBin/../lib/perl5";
-use SneakerNet qw/exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg fullPathToExec/;
+use SneakerNet qw/readTsv exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg fullPathToExec/;
 
-our $VERSION = "1.0";
+our $VERSION = "1.1";
 our $CITATION= "SARS-CoV-2 assembly plugin by Lee Katz.";
 
 local $0=fileparse $0;
@@ -40,7 +41,7 @@ sub main{
       seqtk                           => 'seqtk 2>&1 | grep -m 1 Version:',
       bgzip                           => 'bgzip --version | head -n1',
       tabix                           => 'tabix --version | head -n1',
-      #'v-annotate.pl (VADR)'          => 'v-annotate.pl -h | grep -m 1 [0-9]'
+      'v-annotate.pl (VADR)'          => 'v-annotate.pl -h | grep -m 1 [0-9]'
     }, $settings,
   );
 
@@ -75,6 +76,7 @@ sub assembleAll{
   
   # Find information about each genome
   my $sampleInfo=samplesheetInfo_tsv("$dir/samples.tsv",$settings);
+  my %sampleMetrics = ();
   while(my($sample,$info)=each(%$sampleInfo)){
     next if(ref($info) ne "HASH");
     logmsg "ASSEMBLE SAMPLE $sample";
@@ -99,150 +101,100 @@ sub assembleAll{
         cp($srcFile, $target) or die "ERROR copying $srcFile => $target\n  $!";
       }
       cp($assembly, $outassembly) or die "ERROR copying $assembly => $outassembly\n  $!";
+    }
 
-      mkdir "$outdir/vadr"; # just make this directory right away
+    # Assembly metrics
+    my $asmMetricsFile = "$outdir/assemblyMetrics.tsv";
+    if(! -e $asmMetricsFile){
+      logmsg "Assembly metrics for $outassembly";
+      command("run_assembly_metrics.pl --allMetrics --numcpus $$settings{numcpus} $outassembly >> $asmMetricsFile.tmp");
+      mv("$asmMetricsFile.tmp", $asmMetricsFile) or die "ERROR: could not make $asmMetricsFile";
     }
 
     # Genome annotation
     #logmsg "PREDICT GENES FOR SAMPLE $sample";
-    #if(!-e $outgbk){
-    #  my $gbk=annotateFasta($sample,$outassembly,$settings);
-    #  cp($gbk,$outgbk) or die "ERROR: could not copy $gbk to $outgbk: $!";
-    #}
+    my $vadrPass = "$outdir/vadr/vadr.vadr.pass.tbl";
+    my $vadrFail = "$outdir/vadr/vadr.vadr.fail.tbl";
+    if(!-e $vadrPass || !-e $vadrFail){
+      logmsg "ANNOTATE SAMPLE $sample";
+      my $tmpVadrDir=annotateFasta($sample,$outassembly,$settings);
+      mkdir("$outdir/vadr");
+      for(glob("$tmpVadrDir/*")){
+        next if(!-f $_);
+        cp($_, "$outdir/vadr/".basename($_)) or die "ERROR copying $_ to $outdir/vadr/: $!";
+      }
+    }
[…]
