### 2020-11-02 cd19914
added GP60_Counter v0.7
