+++ scripts/GP60_Counter/Scripts/countGP60repeats.pl
@@ -1,0 +1,209 @@
+#!/usr/bin/perl
+use strict;
+
+#Input is a blast output that contains:
+#
+#qseqid sseqid pident length qcovhsp mismatch gapopen 
+#qstart qend sstart send qlen slen qseq
+#
+#in that order
+#The sseqid needs to be in the format of: Genus_Species_SubtypeFamily_string
+# Ex: C_hominis_IfA12_string
+
+my $input = @ARGV[0]; #tab delim blast output
+
+my $combine;
+
+#For the additional character at the end of
+#subtypes IIc, chipmunk, and skunk
+my $endChar = "";
+
+#For now, we are only going to get the best match so we only want
+#to grab the first line but later we might want to include all
+#matches so keep the loop in and just limit the iterations
+my $loopNum = 0;
+ 
+
+open FILE, $input or die $!;
+while (my $line = <FILE>) {
+
+	$loopNum++;
+	$line =~ s/\r\n/\n/; #remove any Windows characters
+	chomp $line;
+	
+
+	#qseqid sseqid pident length qcovhsp mismatch gapopen qstart qend sstart send qlen slen qseq
+	my ($qseqid, $sseqid, $pident, $length, $qcovhsp, $mismatch, $gapopen, $qstart, $qend, $sstart, $send, $qlen, $slen, $qseq) = split("\t", $line);
+
+	#Check if sequence needs to be reverse complemented
+	#otherwise will not properly count repeats
+	if ($sstart > $send) {
+		my $revcomp = reverse $qseq;
+		$revcomp =~ tr/ATGCatgc/TACGtacg/;
+		$qseq = $revcomp;
+	}
+	
+
+	#parse subtype family from sseqid
+	#sometimes the subtype family will be by itself and sometimes it
+	# will be followed by repeat details always starting with "A"
+	my ($genus, $species, $subFamily, $rest) = split("_", $sseqid, 4);
+	
+	my $subtype;
+	if (index($subFamily, "A") != -1) {
+
+		#parse based on the first occurence of A
+		my ($sub, $rest) = split("A", $subFamily, 2);
+		$subtype = $sub;
+	} else {
+
+		$subtype = $subFamily;
+	}
+
+
+	#Find percentage of Hit matched so we know how much of the
+	#gp60 was found
+	my $hPercent;
+	if ($send > $sstart) {
+
+		$hPercent = (($send-$sstart+1)/$slen)*100;
+	} else {
+
+		$hPercent = (($sstart-$send+1)/$slen)*100;
+	}
+
+	#Ubiquitum reference says no repeats
+	if ($species eq "ubiquitum") {
+
+		print "$input\t$qseqid\t$sseqid\t$subtype\t$length\t$pident\t$hPercent\n";
+	} 
+
+	#IIc, chipmunk, and skunk genotype have an additional
+	#char at the end of their subtypes
+	#>C_parvum_IIcA5G3a_AF164491.1
+	#>Skunk_genotype_XVIa14a_KP099095.1
+	#>Chipmunk_genotypeI_XIVaA18G2T1a_KP099082.1
+	#>C_viatorum_XVaA3a_KP115936.1
[…]
