### 2020-11-02 4b24d18
gp60 counter +x; changed shebang to /usr/bin/env perl
