+++ scripts/GP60_Counter/Scripts/countGP60repeats.pl
@@ -1,4 +1,4 @@
-#!/usr/bin/perl
+#!/usr/bin/env perl
 use strict;
 
 #Input is a blast output that contains:
