### 2020-11-02 3cdab34
v0.16 for crypto; added GP60 software to PATH
