### 2020-11-02 5be930f
formatting GP60 blast database
