### 2020-11-02 357b991
fixed SNVER variable in cd statement
