+++ .github/workflows/unittests.yml
@@ -1,0 +1,46 @@
+on:
+  push:
+    branches: [master, dev]
+
+
+# perl: [ '5.32', '5.30', '5.28' ]
+# perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
+# os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
+# https://github.com/marketplace/actions/setup-perl-environment
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: ['ubuntu-latest' ]
+        #perl: [ '5.16.3' ]
+        perl: [ '5.32', '5.30.1', '5.16.3' ]
+    name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up perl
+        uses: shogo82148/actions-setup-perl@v1
+        with:
+          perl-version: ${{ matrix.perl }}
+      - run: perl -V
+      - name: env
+        run:  env
+      - name: install-makemaker
+        run: |
+          which cpanm
+          cpanm --verbose --notest --force ExtUtils::MakeMaker
+          cpanm --installdeps --verbose --notest --force .
+      - name: perl-check-dependencies
+        run:  |
+          perl Makefile.PL
+          perl -MStatistics::Descriptive -MConfig::Simple -MFile::Slurp -MBio::Kmer -MBio::SeqIO -MText::Fuzzy -MEmail::Stuffer -MBio::FeatureIO -MMoo -MList::MoreUtils -MJSON -e 'print Dumper \%INC'
+          make
+          echo "DEBUG: PERL5LIB: $PERL5LIB"
+      - name: check-mymeta
+        run: |
+          echo "MYMETA.json";
+          cat MYMETA.json
+      - id: prove-lv-t
+        run: |
+          prove -lv t
+
