+++ .github/workflows/unittests.yml
@@ -1,3 +1,4 @@
+name: unit-tests
 on:
   push:
     branches: [master, dev]
