### 2020-11-03 c2fe319
removed custom makefile subroutine
