+++ .github/workflows/unittests.yml
@@ -29,10 +29,12 @@ jobs:
       - name: install-makemaker
         run: |
           which cpanm
-          cpanm --verbose --notest --force ExtUtils::MakeMaker
+          export PERL5LIB=$PERL5LIB:$HOME/lib/perl5
+          cpanm -l $HOME --verbose --notest --force ExtUtils::MakeMaker
           cpanm --installdeps --verbose --notest --force .
       - name: perl-check-dependencies
         run:  |
+          export PERL5LIB=$PERL5LIB:$HOME/lib/perl5
           perl Makefile.PL
           perl -MStatistics::Descriptive -MConfig::Simple -MFile::Slurp -MBio::Kmer -MBio::SeqIO -MText::Fuzzy -MEmail::Stuffer -MBio::FeatureIO -MMoo -MList::MoreUtils -MJSON -e 'print Dumper \%INC'
           make
@@ -43,5 +45,6 @@ jobs:
           cat MYMETA.json
       - id: prove-lv-t
         run: |
+          export PERL5LIB=$PERL5LIB:$HOME/lib/perl5
           prove -lv t
 
