### 2020-11-03 c13f86f
some debugging statements
