+++ .github/workflows/unittests.yml
@@ -30,7 +30,7 @@ jobs:
       - name: tree
         run:  tree
       - name: find LWPSimple
-        run:  find / -name Simple.pm
+        run:  sudo find / -name Simple.pm
       - name: Dumper INC
         run:  perl -MData::Dumper -e 'print \@INC'
       - run: perl -V
