+++ .github/workflows/unittests.yml
@@ -30,7 +30,7 @@ jobs:
       - name: tree
         run:  tree
       - name: Dumper INC
-        run:  perl -MData::Dumper -e 'print \@INC'
+        run:  perl -MData::Dumper -e 'print Dumper \@INC'
       - name: find LWPSimple
         run:  sudo find / -name Simple.pm || true
       - run: perl -V
