+++ .github/workflows/unittests.yml
@@ -24,7 +24,7 @@ jobs:
         with:
           perl-version: ${{ matrix.perl }}
       - name: apt-get install
-        run:  sudo apt-get install tree
+        run:  sudo apt-get install tree libbio-perl-perl libtest-most-perl
       - name: Dumper INC
         run:  perl -MData::Dumper -e 'print Dumper \@INC'
       - run: perl -V
