+++ .github/workflows/unittests.yml
@@ -15,19 +15,18 @@ jobs:
       matrix:
         os: ['ubuntu-latest' ]
         #perl: [ '5.16.3' ]
-        #perl: [ '5.32', '5.30.1', '5.16.3' ]
-        perl: [ '5.32-shrplib', '5.30.1-shrplib', '5.16.3-shrplib' ]
+        perl: [ '5.32', '5.30.1', '5.16.3' ]
     name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v2
-      - name: Set up perl
-        uses: shogo82148/actions-setup-perl@v1
-        with:
-          perl-version: ${{ matrix.perl }}
+      #- uses: actions/checkout@v2
+      #- name: Set up perl
+      #  uses: shogo82148/actions-setup-perl@v1
+      #  with:
+      #    perl-version: ${{ matrix.perl }}
       - name: apt-get install
         run:  |
           sudo apt-get update
-          sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl
+          sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl
       - name: Dumper INC
         run:  perl -MData::Dumper -e 'print Dumper \@INC'
       - run: perl -V
