+++ .github/workflows/unittests.yml
@@ -14,8 +14,8 @@ jobs:
     strategy:
       matrix:
         os: ['ubuntu-latest' ]
-        #perl: [ '5.16.3' ]
-        perl: [ '5.32', '5.30.1', '5.16.3' ]
+        perl: [ '5.16.3' ]
+        #perl: [ '5.32', '5.30.1', '5.16.3' ]
     name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
     steps:
       #- uses: actions/checkout@v2
