+++ .github/workflows/unittests.yml
@@ -58,6 +58,7 @@ jobs:
       - name: perl-check-dependencies
         run:  |
           perl Makefile.PL
+          make
           perl -MStatistics::Descriptive -MConfig::Simple -MFile::Slurp -MBio::Kmer -MBio::SeqIO -MText::Fuzzy -MEmail::Stuffer -MBio::FeatureIO -MMoo -MList::MoreUtils -MJSON -e 'print Dumper \%INC'
           make
           echo "DEBUG: PERL5LIB: $PERL5LIB"
