### 2020-11-03 73a8c4d
touched up installdeps step
