+++ .github/workflows/unittests.yml
@@ -23,10 +23,10 @@ jobs:
       - name: checkout CG-Pipeline
         uses: actions/checkout@v2
           repository: lskatz/CG-Pipeline
-      #- name: Set up perl
-      #  uses: shogo82148/actions-setup-perl@v1
-      #  with:
-      #    perl-version: ${{ matrix.perl }}
+#     - name: Set up perl
+#       uses: shogo82148/actions-setup-perl@v1
+#       with:
+#         perl-version: ${{ matrix.perl }}
       - name: apt-get install
         run:  |
           sudo apt-get update
@@ -38,21 +38,21 @@ jobs:
       - run: perl -V
       - name: env
         run:  env
-      #- name: install Bio::SeqIO
-      #  run: |
-      #    export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-      #    which cpanm
-      #    sudo cpanm --verbose --notest --force Bio::SeqIO 
-      #- name: install Bio::FeatureIO
-      #  run: |
-      #    export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-      #    which cpanm
-      #    sudo cpanm --verbose --notest --force Bio::FeatureIO 
-      #- name: install Bio::Kmer
-      #  run: |
-      #    export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-      #    which cpanm
-      #    sudo cpanm --verbose --notest --force Bio::Kmer 
+#     - name: install Bio::SeqIO
+#       run: |
+#         export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
+#         which cpanm
+#         sudo cpanm --verbose --notest --force Bio::SeqIO 
+#     - name: install Bio::FeatureIO
+#       run: |
+#         export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
+#         which cpanm
+#         sudo cpanm --verbose --notest --force Bio::FeatureIO 
+#     - name: install Bio::Kmer
+#       run: |
+#         export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
+#         which cpanm
+#         sudo cpanm --verbose --notest --force Bio::Kmer 
       - name: install-makemaker
         run: |
           which cpanm
