+++ .github/workflows/unittests.yml
@@ -8,6 +8,11 @@ on:
 # perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
 # os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
 # https://github.com/marketplace/actions/setup-perl-environment
+#     - name: Set up perl
+#       uses: shogo82148/actions-setup-perl@v1
+#       with:
+#         perl-version: ${{ matrix.perl }}
+
 jobs:
   build:
     runs-on: ${{ matrix.os }}
@@ -23,10 +28,6 @@ jobs:
       - name: checkout CG-Pipeline
         uses: actions/checkout@v2
           repository: lskatz/CG-Pipeline
-#     - name: Set up perl
-#       uses: shogo82148/actions-setup-perl@v1
-#       with:
-#         perl-version: ${{ matrix.perl }}
       - name: apt-get install
         run:  |
           sudo apt-get update
