+++ .github/workflows/unittests.yml
@@ -11,6 +11,7 @@ on:
 #       uses: shogo82148/actions-setup-perl@v1
 #       with:
 #         perl-version: ${{ matrix.perl }}
+#   name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
 
 jobs:
   build:
@@ -20,7 +21,6 @@ jobs:
         os: ['ubuntu-latest' ]
         perl: [ '5.30.1' ]
         #perl: [ '5.32', '5.30.1', '5.16.3' ]
-#   name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
    name: System Perl on ${{ matrix.os }}
     steps:
       - name: checkout my repo
