+++ .github/workflows/unittests.yml
@@ -19,9 +19,7 @@ jobs:
     strategy:
       matrix:
         os: ['ubuntu-latest' ]
-        perl: [ '5.30.1' ]
-        #perl: [ '5.32', '5.30.1', '5.16.3' ]
-   name: System Perl on ${{ matrix.os }}
+  name: System Perl on ${{ matrix.os }}
     steps:
       - name: checkout my repo
         uses: actions/checkout@v2
