### 2020-11-03 aceb6d1
bring back manual dep installation
