### 2020-11-03 b1a5968
bring back manual dep installation
