### 2020-11-03 c0aa268
responsibly put CG-Pipeline into a subfolder
