+++ .github/workflows/unittests.yml
@@ -27,6 +27,7 @@ jobs:
         uses: actions/checkout@v2
         with:
           repository: lskatz/CG-Pipeline
+          path: CG-Pipeline
       - name: apt-get install
         run:  |
           sudo apt-get update
