+++ .github/workflows/unittests.yml
@@ -32,6 +32,18 @@ jobs:
         run:  |
           sudo apt-get update
           sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl cpanminus
+          sudo apt-get -y install build-essential bioperl-run ncbi-blast+ prodigal sendmail python3 python3-pip
+      - name: update alternatives for python3
+        run:  |
+          sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 1
+          sudo update-alternatives --set python /usr/bin/python3
+          python --version
+          pip3 --version
+      - name: configure cpanm
+        run:  |
+          cpanm --local-lib=$HOME/perl5 local::lib
+          perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib
+          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
       - name: Dumper INC
         run:  perl -MData::Dumper -e 'print Dumper \@INC'
       - name: tree
@@ -41,7 +53,7 @@ jobs:
         run:  env
       - name: install Bio::SeqIO
         run: |
-          export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
+          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           which cpanm
           sudo cpanm --verbose --notest --force Bio::SeqIO 
       - name: install Bio::FeatureIO
