### 2020-11-03 7c84b34
removed some sudo steps; tried to do the right cpanm commands
