+++ .github/workflows/unittests.yml
@@ -51,42 +51,41 @@ jobs:
       - run: perl -V
       - name: env
         run:  env
-      - name: install Bio::SeqIO
-        run: |
-          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
-          which cpanm
-          sudo cpanm --verbose --notest --force Bio::SeqIO 
-      - name: install Bio::FeatureIO
-        run: |
-          export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-          which cpanm
-          sudo cpanm --verbose --notest --force Bio::FeatureIO 
+#     - name: install Bio::SeqIO
+#       run: |
+#         eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+#         which cpanm
+#         sudo cpanm --verbose --notest --force Bio::SeqIO 
+#     - name: install Bio::FeatureIO
+#       run: |
+#         export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
+#         which cpanm
+#         sudo cpanm --verbose --notest --force Bio::FeatureIO 
       - name: install Bio::Kmer
         run: |
-          export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-          which cpanm
-          sudo cpanm --verbose --notest --force Bio::Kmer 
+          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+          cpanm -l $HOME/perl5 --verbose --notest --force Bio::Kmer 
       - name: install-makemaker
         run: |
-          which cpanm
-          export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-          sudo cpanm --verbose --notest --force ExtUtils::MakeMaker
-          sudo cpanm --installdeps -l . --verbose --notest --force
+          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+          cpanm -l $HOME/perl5 --verbose --notest --force ExtUtils::MakeMaker
+          cpanm -l $HOME/perl5 --installdeps -l . --verbose --notest --force
       - name: perl-check-dependencies
         run:  |
           export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
-          sudo perl Makefile.PL
-          sudo make
-          sudo cpanm --installdeps . -l . --verbose --notest --force
+          perl Makefile.PL
+          make
+          cpanm --installdeps . -l . --verbose --notest --force
           perl -I$(pwd)/lib/perl5 -MData::Dumper -MStatistics::Descriptive -MConfig::Simple -MFile::Slurp -MBio::Kmer -MBio::SeqIO -MText::Fuzzy -MEmail::Stuffer -MBio::FeatureIO -MMoo -MList::MoreUtils -MJSON -e 'print Dumper \%INC'
-          sudo make
+          make
           echo "DEBUG: PERL5LIB: $PERL5LIB"
       - name: check-mymeta
         run: |
           echo "MYMETA.json";
-          sudo cat MYMETA.json
-      - id: prove-lv-t
-        run: |
-          export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
+          cat MYMETA.json
+      - id:   prove-lv-t
+        name: prove-lv-t
+        run:  |
+          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           sudo prove -lv t
 
