+++ .github/workflows/unittests.yml
@@ -69,13 +69,13 @@ jobs:
         run: |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           cpanm -l $HOME/perl5 --verbose --notest --force ExtUtils::MakeMaker
-          cpanm -l $HOME/perl5 --installdeps -l . --verbose --notest --force
+          cpanm -l $HOME/perl5 --installdeps . --verbose --notest --force
       - name: perl-check-dependencies
         run:  |
           export PERL5LIB=$PERL5LIB:$(pwd)/lib/perl5
           perl Makefile.PL
           make
-          cpanm --installdeps . -l . --verbose --notest --force
+          cpanm --installdeps . -l $HOME/perl5 --verbose --notest --force
           perl -I$(pwd)/lib/perl5 -MData::Dumper -MStatistics::Descriptive -MConfig::Simple -MFile::Slurp -MBio::Kmer -MBio::SeqIO -MText::Fuzzy -MEmail::Stuffer -MBio::FeatureIO -MMoo -MList::MoreUtils -MJSON -e 'print Dumper \%INC'
           make
           echo "DEBUG: PERL5LIB: $PERL5LIB"
