### 2020-11-03 bc89494
fixed more PERL5 env vars
