+++ .github/workflows/unittests.yml
@@ -87,5 +87,5 @@ jobs:
         name: prove-lv-t
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
-          sudo prove -lv t
+          prove -lv t
 
