### 2020-11-03 0916a24
removed specific Bio::Kmer installation; added a debugging statement for env
