+++ .github/workflows/unittests.yml
@@ -96,9 +96,12 @@ jobs:
         name: prove-lv-t
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
-          export PATH=$PATH:~/CG-Pipeline/scripts
-          export PATH=$PATH:~/shovill-1.0.0/bin
-          export PATH=$PATH:~/.local/bin
-          export PATH=$PATH:~/krona
+          export PATH=$PATH:$HOME/CG-Pipeline/scripts
+          export PATH=$PATH:$HOME/shovill-1.0.0/bin
+          export PATH=$PATH:$HOME/.local/bin
+          export PATH=$PATH:$HOME/krona
+          echo
+          echo "$PATH" | tr ":" "\n"
+          echo
           prove -lv t
 
