### 2020-11-03 c0d86e6
fixed PATH with GITHUB_WORKSPACE; added more tree debugging
