+++ .github/workflows/unittests.yml
@@ -59,8 +59,12 @@ jobs:
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
       - name: Dumper INC
         run:  perl -MData::Dumper -e 'print Dumper \@INC'
-      - name: tree
-        run: tree
+      - name: tree pwd
+        run: tree $(pwd)
+      - name: tree home
+        run: tree $HOME
+      - name: tree GITHUB_WORKSPACE
+        run: tree $GITHUB_WORKSPACE
       - run: perl -V
       - name: env
         run:  env
@@ -96,10 +100,10 @@ jobs:
         name: prove-lv-t
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
-          export PATH=$PATH:$HOME/CG-Pipeline/scripts
-          export PATH=$PATH:$HOME/shovill-1.0.0/bin
-          export PATH=$PATH:$HOME/.local/bin
-          export PATH=$PATH:$HOME/krona
+          export PATH=$PATH:$GITHUB_WORKSPACE/CG-Pipeline/scripts
+          export PATH=$PATH:$GITHUB_WORKSPACE/shovill-1.0.0/bin
+          export PATH=$PATH:$GITHUB_WORKSPACE/.local/bin
+          export PATH=$PATH:$GITHUB_WORKSPACE/krona
           echo
           echo "$PATH" | tr ":" "\n"
           echo
