+++ .github/workflows/unittests.yml
@@ -3,6 +3,8 @@ on:
   push:
     branches: [master, dev, unit-test]
 
+env:
+  spadesVer: 3.14.1
 # perl: [ '5.32', '5.30', '5.28' ]
 # perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
 # os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
@@ -46,12 +48,18 @@ jobs:
           sudo apt-get update
           sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl cpanminus
           sudo apt-get -y install build-essential bioperl-run ncbi-blast+ prodigal sendmail python3 python3-pip
+          # shovill things
+          sudo apt-get -y install bwa velvet wget pigz openjdk-9-jre libncurses5-dev libbz2-dev liblzma-dev libfindbin-libs-perl trimmomatic
       - name: update alternatives for python3
         run:  |
           sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 1
           sudo update-alternatives --set python /usr/bin/python3
           python --version
           pip3 --version
+      - name: install SPAdes
+        run:  |
+          wget https://github.com/ablab/spades/releases/download/v${spadesVer}/SPAdes-${spadesVer}-Linux.tar.gz
+          tar -xzf SPAdes-${spadesVer}-Linux.tar.gz
       - name: configure cpanm
         run:  |
           cpanm --local-lib=$HOME/perl5 local::lib
@@ -104,6 +112,9 @@ jobs:
           export PATH=$PATH:$GITHUB_WORKSPACE/shovill-1.0.0/bin
           export PATH=$PATH:$GITHUB_WORKSPACE/.local/bin
           export PATH=$PATH:$GITHUB_WORKSPACE/krona
+          export PATH=$PATH:$GITHUB_WORKSPACE/SPAdes-${spadesVer}-Linux/bin
+          export PATH=$PATH:$GITHUB_WORKSPACE/kmc
+          export PATH=$PATH:$GITHUB_WORKSPACE/Lighter-1.1.1
           echo
           echo "$PATH" | tr ":" "\n"
           echo
