### 2020-11-04 9aab61e
removed openjdk-9-jre from apt-get install
