+++ .github/workflows/unittests.yml
@@ -49,7 +49,7 @@ jobs:
           sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl cpanminus
           sudo apt-get -y install build-essential bioperl-run ncbi-blast+ prodigal sendmail python3 python3-pip
           # shovill things
-          sudo apt-get -y install bwa velvet wget pigz openjdk-9-jre libncurses5-dev libbz2-dev liblzma-dev libfindbin-libs-perl trimmomatic
+          sudo apt-get -y install bwa velvet wget pigz libncurses5-dev libbz2-dev liblzma-dev libfindbin-libs-perl trimmomatic
       - name: update alternatives for python3
         run:  |
           sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 1
