### 2020-11-04 c3f1d5c
installing more shovil prereqs; debugging test 90 more
