### 2020-11-04 eb7dcbf
more debugging statements
