+++ .github/workflows/unittests.yml
@@ -66,6 +66,14 @@ jobs:
           sudo ln -sv $GITHUB_WORKSPACE/kmc/kmc* /usr/local/bin
           which kmc
           kmc --help || true
+      - name: install lighter
+        run:  |
+          mkdir -v $GITHUB_WORKSPACE/lighter
+          cd $GITHUB_WORKSPACE/lighter
+          wget https://github.com/mourisl/Lighter/archive/v1.1.1.tar.gz
+          tar -zvxf v1.1.1.tar.gz
+          make
+          sudo ln -sv lighter /usr/local/bin
       - name: apt-get install
         run:  |
           sudo apt-get update
@@ -114,6 +122,7 @@ jobs:
       - name: install-makemaker
         run: |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+          cd SneakerNet
           cpanm -l $HOME/perl5 --verbose --notest --force ExtUtils::MakeMaker
           cpanm -l $HOME/perl5 --installdeps . --verbose --notest --force
       - name: perl-check-dependencies
