### 2020-11-07 c810e21
cd $GITHUB_WORKSPACE/SneakerNet
