### 2020-11-07 2750ff3
cd $GITHUB_WORKSPACE/SneakerNet
