### 2020-11-07 d8bfabd
install trimmomatic
