+++ .github/workflows/unittests.yml
@@ -9,10 +9,6 @@ env:
 # perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
 # os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
 # https://github.com/marketplace/actions/setup-perl-environment
-#     - name: Set up perl
-#       uses: shogo82148/actions-setup-perl@v1
-#       with:
-#         perl-version: ${{ matrix.perl }}
 #   name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
 
 jobs:
@@ -21,8 +17,14 @@ jobs:
     strategy:
       matrix:
         os: ['ubuntu-latest' ]
+        perl: [ '5.32' ]
     name: System Perl on ${{ matrix.os }}
     steps:
+      - name: Set up perl
+        uses: shogo82148/actions-setup-perl@v1
+        with:
+          perl-version: ${{ matrix.perl }}
+          multi-thread: "true"
       - name: checkout my repo
         uses: actions/checkout@v2
         with:
