+++ .github/workflows/unittests.yml
@@ -9,7 +9,6 @@ env:
 # perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
 # os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
 # https://github.com/marketplace/actions/setup-perl-environment
-#   name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
 
 jobs:
   build:
@@ -18,7 +17,7 @@ jobs:
       matrix:
         os: ['ubuntu-latest' ]
         perl: [ '5.32' ]
-    name: System Perl on ${{ matrix.os }}
+    name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
     steps:
       - name: Set up perl
         uses: shogo82148/actions-setup-perl@v1
