+++ .github/workflows/unittests.yml
@@ -80,7 +80,7 @@ jobs:
       - name: install trimmomatic
         run:  |
           mkdir $GITHUB_WORKSPACE/trimmomatic 
-          cd $GITHUB_WORKSPACE trimmomatic
+          cd $GITHUB_WORKSPACE/trimmomatic
           wget http://www.usadellab.org/cms/uploads/supplementary/Trimmomatic/Trimmomatic-0.38.zip 
           unzip Trimmomatic-0.38.zip 
           rm -rf Trimmomatic-0.38.zip 
