### 2020-11-10 2167309
install mash v2.2
