+++ .github/workflows/unittests.yml
@@ -62,7 +62,7 @@ jobs:
           mkdir $GITHUB_WORKSPACE/Mash-2.2
           cd $GITHUB_WORKSPACE/Mash-2.2
           wget https://github.com/marbl/Mash/releases/download/v2.2/mash-Linux64-v2.2.tar
-          untar -vf mash-Linux64-v2.2.tar
+          tar -xvf mash-Linux64-v2.2.tar
           ln -sv mash-Linux64-v2.2 mash
           export PATH=$PATH:$GITHUB_WORKSPACE/Mash-2.2
           echo "$GITHUB_WORKSPACE/Mash-2.2" >> $GITHUB_PATH
