+++ .github/workflows/unittests.yml
@@ -59,13 +59,12 @@ jobs:
           ref:  v1.1.0
       - name: install Mash
         run:  |
-          mkdir $GITHUB_WORKSPACE/Mash-2.2
-          cd $GITHUB_WORKSPACE/Mash-2.2
+          cd $GITHUB_WORKSPACE
           wget https://github.com/marbl/Mash/releases/download/v2.2/mash-Linux64-v2.2.tar
-          tar -xvf mash-Linux64-v2.2.tar
+          tar -xvf mash-Linux64-v2.2.tar # creates mash-Linux64-v2.2/mash
           ln -sv mash-Linux64-v2.2 mash
-          export PATH=$PATH:$GITHUB_WORKSPACE/Mash-2.2
-          echo "$GITHUB_WORKSPACE/Mash-2.2" >> $GITHUB_PATH
+          export PATH=$PATH:$GITHUB_WORKSPACE/mash-Linux64-v2.2
+          echo "$GITHUB_WORKSPACE/mash-Linux64-v2.2" >> $GITHUB_PATH
           which mash
       - name: checkout seqtk
         run:  |
