+++ .github/workflows/unittests.yml
@@ -105,6 +105,7 @@ jobs:
           make
           chmod -v +x lighter
           sudo ln -sv lighter /usr/local/bin
+          which lighter
       - name: install trimmomatic
         run:  |
           mkdir $GITHUB_WORKSPACE/trimmomatic 
@@ -117,6 +118,7 @@ jobs:
           echo "exec java -jar $GITHUB_WORKSPACE/trimmomatic/Trimmomatic-0.38/trimmomatic-0.38.jar """"$""@"""" " >> trimmomatic
           chmod -v +x trimmomatic
           sudo ln -sv trimmomatic /usr/local/bin/
+          which trimmomatic
       - name: apt-get install
         run:  |
           sudo apt-get update
@@ -146,6 +148,9 @@ jobs:
           mv kmercounter.centos7.7 kmercounter
           chmod +x skesa gfa_connector kmercounter
           sudo ln -sv skesa gfa_connector kmercounter /usr/local/bin
+          which skesa
+          which gfa_connector
+          which kmercounter
       - name: configure cpanm
         run:  |
           cpanm --local-lib=$HOME/perl5 local::lib
@@ -200,8 +205,10 @@ jobs:
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           echo
+          echo "PATH broken up:"
           echo "$PATH" | tr ":" "\n"
           echo
+          which flash
           cd $GITHUB_WORKSPACE/SneakerNet
           prove -lv t
 
