### 2020-11-12 10989b3
tree command on flash directory
