### 2020-11-12 d58f96b
recursive ls instead of tree for flash
