+++ .github/workflows/unittests.yml
@@ -107,7 +107,8 @@ jobs:
           cd Lighter-1.1.1
           make
           chmod -v +x lighter
-          sudo ln -sv lighter /usr/local/bin
+          export $PATH=$PATH:$GITHUB_WORKSPACE/lighter
+          echo "$GITHUB_WORKSPACE/lighter" >> $GITHUB_PATH
           which lighter
       - name: install trimmomatic
         run:  |
