### 2020-11-12 9fadbdc
fixed syntax error for lighter installation
