+++ .github/workflows/unittests.yml
@@ -107,7 +107,7 @@ jobs:
           make
           chmod -v +x lighter
           export PATH=$PATH:$GITHUB_WORKSPACE/lighter-1.1.1
-          echo "$GITHUB_WORKSPACE/lighter" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/lighter-1.1.1" >> $GITHUB_PATH
           which lighter
       - name: install megahit
         run:  |
