### 2020-11-12 2447283
fix trimmomatic path
