+++ .github/workflows/unittests.yml
@@ -117,18 +117,20 @@ jobs:
           tar -xzf megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin.tar.gz
           export PATH="$PATH:$GITHUB_WORKSPACE/megahit"
           echo "$GITHUB_WORKSPACE/megahit" >> $GITHUB_PATH
+          which megahit
       - name: install trimmomatic
         run:  |
-          mkdir $GITHUB_WORKSPACE/trimmomatic 
-          cd $GITHUB_WORKSPACE/trimmomatic
+          cd $GITHUB_WORKSPACE
           wget http://www.usadellab.org/cms/uploads/supplementary/Trimmomatic/Trimmomatic-0.38.zip 
           unzip Trimmomatic-0.38.zip 
           rm -rf Trimmomatic-0.38.zip 
-          chmod +x Trimmomatic-0.38/trimmomatic-0.38.jar
+          cd Trimmomatic-0.38
+          chmod +x trimmomatic-0.38.jar
           echo "#!/bin/bash" >> trimmomatic
-          echo "exec java -jar $GITHUB_WORKSPACE/trimmomatic/Trimmomatic-0.38/trimmomatic-0.38.jar """"$""@"""" " >> trimmomatic
+          echo "exec java -jar $GITHUB_WORKSPACE/Trimmomatic-0.38/trimmomatic-0.38.jar """"$""@"""" " >> trimmomatic
           chmod -v +x trimmomatic
-          sudo ln -sv trimmomatic /usr/local/bin/
+          export PATH="$PATH:$GITHUB_WORKSPACE/Trimmomatic-0.38"
+          echo "$GITHUB_WORKSPACE/Trimmomatic-0.38" >> $GITHUB_PATH
           which trimmomatic
       - name: apt-get install
         run:  |
