### 2020-11-12 1b70bfd
ls recursive for megahit
