+++ .github/workflows/unittests.yml
@@ -115,6 +115,7 @@ jobs:
           cd "$GITHUB_WORKSPACE/megahit"
           wget https://github.com/voutcn/megahit/releases/download/v1.1.4/megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin.tar.gz
           tar -xzf megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin.tar.gz
+          ls -lhSR "$GITHUB_WORKSPACE/megahit"
           export PATH="$PATH:$GITHUB_WORKSPACE/megahit"
           echo "$GITHUB_WORKSPACE/megahit" >> $GITHUB_PATH
           which megahit
