### 2020-11-12 3eaef3d
added subdir for megahit bin path
