+++ .github/workflows/unittests.yml
@@ -116,8 +116,8 @@ jobs:
           wget https://github.com/voutcn/megahit/releases/download/v1.1.4/megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin.tar.gz
           tar -xzf megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin.tar.gz
           ls -lhSR "$GITHUB_WORKSPACE/megahit"
-          export PATH="$PATH:$GITHUB_WORKSPACE/megahit"
-          echo "$GITHUB_WORKSPACE/megahit" >> $GITHUB_PATH
+          export PATH="$PATH:$GITHUB_WORKSPACE/megahit/megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin"
+          echo "$GITHUB_WORKSPACE/megahit/megahit_v1.1.4_LINUX_CPUONLY_x86_64-bin" >> $GITHUB_PATH
           which megahit
       - name: install trimmomatic
         run:  |
