### 2020-11-12 8fbe88c
samclip installation
