### 2020-11-12 74bc5ef
samclip installation
