### 2020-11-12 7a0c517
jellyfish installation esp for Kraken
