### 2020-11-12 3c4a58b
install kraken1
