+++ .github/workflows/unittests.yml
@@ -190,6 +190,13 @@ jobs:
           export PATH=$PATH:/opt/kraken
           echo "/opt/kraken/" >> $GITHUB_PATH
           which kraken
+      - name: install minikraken
+        run: |
+          sudo mkdir -v /kraken-database
+          cd /kraken-database
+          wget https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz
+          tar -zxf minikraken_20171019_4GB.tgz 
+          ls -lhSR /kraken-database
       - name: install SPAdes
         run:  |
           wget https://github.com/ablab/spades/releases/download/v${spadesVer}/SPAdes-${spadesVer}-Linux.tar.gz
