### 2020-11-12 2353bf8
fix kraken installation
