+++ .github/workflows/unittests.yml
@@ -183,7 +183,7 @@ jobs:
         run:  |
           cd $GITHUB_WORKSPACE
           wget https://github.com/DerrickWood/kraken/archive/v1.1.1.tar.gz -O kraken-v1.1.1.tar.gz
-          tar -xvzf kraken-v1.1.1.tar.gzv1.1.1.tar.gz 
+          tar -xvzf kraken-v1.1.1.tar.gz
           cd kraken-1.1.1
           mkdir /opt/kraken
           ./install_kraken.sh /opt/kraken/
