+++ .github/workflows/unittests.yml
@@ -194,9 +194,9 @@ jobs:
         run: |
           sudo mkdir -v /kraken-database
           cd /kraken-database
-          wget https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz
-          tar -zxf minikraken_20171019_4GB.tgz 
-          ls -lhSR /kraken-database
+          sudo wget https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz
+          sudo tar -zxf minikraken_20171019_4GB.tgz 
+          sudo ls -lhSR /kraken-database
       - name: install SPAdes
         run:  |
           wget https://github.com/ablab/spades/releases/download/v${spadesVer}/SPAdes-${spadesVer}-Linux.tar.gz
