+++ .github/workflows/unittests.yml
@@ -5,6 +5,7 @@ on:
 
 env:
   spadesVer: 3.14.1
+  KRAKEN_DEFAULT_DB: /kraken-database
 # perl: [ '5.32', '5.30', '5.28' ]
 # perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
 # os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
@@ -250,6 +251,11 @@ jobs:
       - name: install-makemaker
         run: |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+          perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib >> $GITHUB_ENV
+          echo 
+          echo GITHUG_ENV
+          cat $GITHUB_ENV
+          echo
           cd SneakerNet
           cpanm -l $HOME/perl5 --verbose --notest --force ExtUtils::MakeMaker
           cpanm -l $HOME/perl5 --installdeps . --verbose --notest --force
@@ -277,5 +283,5 @@ jobs:
           echo
           which flash
           cd $GITHUB_WORKSPACE/SneakerNet
-          prove -lv t
+          sudo prove -lv t
 
