### 2020-11-12 9d17541
SneakerNet debug statements
