+++ .github/workflows/unittests.yml
@@ -269,19 +269,22 @@ jobs:
           perl -MData::Dumper -MStatistics::Descriptive -MConfig::Simple -MFile::Slurp -MBio::Kmer -MBio::SeqIO -MText::Fuzzy -MEmail::Stuffer -MBio::FeatureIO -MMoo -MList::MoreUtils -MJSON -e 'print Dumper \%INC'
           make
           echo "DEBUG: PERL5LIB: $PERL5LIB"
-      - name: check-mymeta
+      - name: check-SneakerNet-repo
         run: |
           cd $GITHUB_WORKSPACE/SneakerNet
           cat MYMETA.json
-      - id:   prove-lv-t
-        name: prove-lv-t
-        run:  |
-          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+          echo;env;echo
+          ls -lhS t/M00123-18-001-test/
+          md5sum t/M00123-18-001-test/*.fastq.gz
           echo
           echo "PATH broken up:"
           echo "$PATH" | tr ":" "\n"
           echo
           which flash
+      - id:   prove-lv-t
+        name: prove-lv-t
+        run:  |
+          eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           cd $GITHUB_WORKSPACE/SneakerNet
           sudo prove -lv t
 
