+++ .github/workflows/unittests.yml
@@ -291,5 +291,5 @@ jobs:
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           cd $GITHUB_WORKSPACE/SneakerNet
-          sudo prove -lv t
+          prove -lv t
 
