+++ .github/workflows/unittests.yml
@@ -281,15 +281,23 @@ jobs:
           echo "PATH broken up:"
           echo "$PATH" | tr ":" "\n"
           echo
-          echo "GITHUB_ENV"
+          echo "GITHUB_ENV: $GITHUB_ENV"
           cat $GITHUB_ENV
-          echo "GITHUB_PATH"
+          echo "GITHUB_PATH: $GITHUB_PATH"
           cat $GITHUB_PATH
+          echo "Which flash:"
           which flash
+          echo
+          echo "Running t/00_env.t ..."
+          perl t/00_env.t
+          for i in t/M00123-18-001-test/*.fastq.gz; do echo -ne "$i\t"; zcat $i | tail -n 1; done;
       - id:   prove-lv-t
         name: prove-lv-t
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
           cd $GITHUB_WORKSPACE/SneakerNet
-          prove -lv t
+          #prove -lv t
+          make
+          for i in t/*.t; do perl t; done;
+
 
