### 2020-11-12 a8c0fdb
remove export statements when printing to GITHUB_ENV
