### 2020-11-12 2b83d2f
debugging broken pipe for gzip
