+++ .github/workflows/unittests.yml
@@ -299,6 +299,8 @@ jobs:
         name: prove-lv-t
         run:  |
           eval $(perl -I $HOME/perl5/lib/perl5/ -Mlocal::lib)
+          mkdir -v $GITHUB_WORKSPACE/tmp
+          export TMPDIR=$GITHUB_WORKSPACE/tmp
           cd $GITHUB_WORKSPACE/SneakerNet
           #prove -lv t
           make
