### 2020-11-12 97d368b
skipping asseembly tests when in CI environment
