+++ .github/workflows/unittests.yml
@@ -304,6 +304,6 @@ jobs:
           cd $GITHUB_WORKSPACE/SneakerNet
           #prove -lv t
           make
-          for i in t/*.t; do perl $i; done;
+          make test
 
 
