### 2020-11-12 9014c8b
adding correct place for kraken db
