### 2020-11-12 1ca6cbf
fixing installation of krona
