### 2020-11-12 aec6770
fixing /kraken-database folder
