### 2020-11-12 bcd2d5e
GitHub Actions unit testing (#55)
* some debugging statements

* more debugging

* more debugging INC

* find >dev/null

* push find command further down

* push find command further down

* fix find command to /opt/hostedtoolcache

* separately install Bio::* modules

* test bioperl

* cpanm to PWD

* added dependency packages listed on packages.ubuntu.com/bionic/bioperl

* apt-get update

* perl5lib update for bioperl/cpanm installation

* perl5lib update for bioperl/cpanm installation; fixed $pwd

* removed no-install-recommends; separate out cpanm steps

* berkeleydb

* libdb-dev

* trying
[…]
