+++ Makefile.PL
@@ -60,7 +60,7 @@ WriteMakefile(
     'Thread::Queue'  => 0,
     # Not core (I think?)
     'Statistics::Descriptive' => 0,
-    'Config::Simple' => 0,
+    'Config::Simple' => '4.58',
     'File::Slurp'    => 0,
     'Bio::Kmer'      => 0,
     'Bio::SeqIO'     => 0,
