+++ .github/workflows/unittests.yml
@@ -1,50 +1,309 @@
 name: unit-tests
 on:
   push:
-    branches: [master, dev]
+    branches: [master, dev, unit-test]
 
-
+env:
+  spadesVer: 3.14.1
+  KRAKEN_DEFAULT_DB: /kraken-database
 # perl: [ '5.32', '5.30', '5.28' ]
 # perl: [ '5.32', '5.30.1', '5.16.3', '5.12.3' ]
 # os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
 # https://github.com/marketplace/actions/setup-perl-environment
+
 jobs:
   build:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         os: ['ubuntu-latest' ]
-        #perl: [ '5.16.3' ]
-        perl: [ '5.32', '5.30.1', '5.16.3' ]
+        perl: [ '5.32' ]
     name: Perl ${{ matrix.perl }} on ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v2
+      - name: add some paths to PATH
+        run:  |
+          echo "$GITHUB_WORKSPACE/CG-Pipeline/scripts" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/shovill/bin" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/.local/bin" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/krona" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/SPAdes-${spadesVer}-Linux/bin" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/kmc" >> $GITHUB_PATH
+          echo "$GITHUB_WORKSPACE/Lighter-1.1.1" >> $GITHUB_PATH
+          echo
+          cat $GITHUB_PATH
       - name: Set up perl
         uses: shogo82148/actions-setup-perl@v1
         with:
           perl-version: ${{ matrix.perl }}
+          multi-thread: "true"
+      - name: checkout my repo
+        uses: actions/checkout@v2
+        with:
+          path: SneakerNet
+      - name: checkout CG-Pipeline
+        uses: actions/checkout@v2
+        with:
+          repository: lskatz/CG-Pipeline
+          path: CG-Pipeline
+          ref:  v0.5
+      - name: checkout Shovill
+        uses: actions/checkout@v2
+        with:
+          repository: tseemann/shovill
+          path: shovill
+          ref:  v1.1.0
+      - name: install krona
+        run:  |
+          cd $GITHUB_WORKSPACE
+          wget https://github.com/marbl/Krona/releases/download/v2.7.1/KronaTools-2.7.1.tar
+          tar xvf KronaTools-2.7.1.tar
+          cd KronaTools-2.7.1
+          ./install.pl --prefix /opt/krona
+          ./updateTaxonomy.sh
+          export PATH=$PATH:/opt/krona/bin
+          echo "/opt/krona/bin" >> $GITHUB_PATH
+          which ktImportText
+      - name: install samclip
+        run:  |
+          mkdir $GITHUB_WORKSPACE/samclip
+          cd $GITHUB_WORKSPACE/samclip
+          wget https://raw.githubusercontent.com/tseemann/samclip/master/samclip
+          chmod +x samclip
+          export PATH=$PATH:$GITHUB_WORKSPACE/samclip
+          echo "$GITHUB_WORKSPACE/samclip" >> $GITHUB_PATH
+          which samclip
+      - name: install FLASh
+        run:  |
+          cd $GITHUB_WORKSPACE
[…]
