### 2020-11-16 #2334 @erinyoung
I still cannot get the test to work. I was hoping that our MiSeqs would be hooked up so I could test it as directly as suggested by the documentation.

The last thing that I tried (today)

I pulled a new docker image and ran the following:
```
$ docker run --rm -v $PWD:/data -v $KRAKEN_DEFAULT_DB:/kraken-database -u $(id -u):$(id -g) lskatz/sneakernet:latest
```
And the following was printed to my screen:
```
SneakerNetPlugins.pl --numcpus 8 t/M00123-18-001-test
sn_parseSampleSheet.pl: Found ./samples.tsv.  Not reparsing without --force.
addReadMetrics.pl: Tempdir is /tmp/addReadMetrics.pl._bgSBL
assembleAll.pl: Temporary directory is at /tmp/assembleAll.pl.HVx6A5
assembleAll.pl: ASSEMBLE SAMPLE LT2
assembleAll.pl: PREDICT GENES FOR SAMPLE LT2
assembleAll.pl: ASSEMBLE SAMPLE Philadelphia_CDC
assembleAll.pl: PREDICT GENES FOR SAMPLE Philadelphia_CDC
assembleAll.pl: ASSEMBLE SAMPLE 2010EL-1786
assembleAll.pl: PREDICT GENES FOR SAMPLE 2010EL-1786
assembleAll.pl: ASSEMBLE SAMPLE FA1090
assembleAll.pl: PREDICT GENES FOR SAMPLE FA1090
assembleAll.pl: ASSEMBLE SAMPLE contaminated
assembleAll.pl: PREDICT GENES FOR SAMPLE contaminated
assembleAll.pl: Running metrics on the genbank files at ./SneakerNet/forEmail/assemblyMetrics.tsv
assembleAll.pl: gbk metrics for ./SneakerNet/assemblies/2010EL-1786/2010EL-1786.shovill.skesa.gbk
assembleAll.pl: gbk metrics for ./SneakerNet/assemblies/contaminated/contaminated.shovill.skesa.gbk
assembleAll.pl: gbk metrics for ./SneakerNet/assemblies/FA1090/FA1090.shovill.skesa.gbk
assembleAll.pl: gbk metrics for ./SneakerNet/assemblies/LT2/LT2.shovill.skesa.gbk
assembleAll.pl: gbk metrics for ./SneakerNet/assemblies/Philadelphia_CDC/Philadelphia_CDC.shovill.skesa.gbk
assembleAll.pl: asm metrics for ./SneakerNet/assemblies/FA1090/FA1090.shovill.skesa.fasta
assembleAll.pl: asm metrics for ./SneakerNet/assemblies/2010EL-1786/2010EL-1786.shovill.skesa.fasta
assembleAll.pl: asm metrics for ./SneakerNet/assemblies/Philadelphia_CDC/Philadelphia_CDC.shovill.skesa.fasta
assembleAll.pl: asm metrics for ./SneakerNet/assemblies/contaminated/contaminated.shovill.skesa.fasta
assembleAll.pl: asm metrics for ./SneakerNet/assemblies/LT2/LT2.shovill.skesa.fasta
assembleAll.pl: Effective coverage for ./SneakerNet/assemblies/2010EL-1786/2010EL-1786.shovill.skesa.bam
assembleAll.pl: Effective coverage for ./SneakerNet/assemblies/FA1090/FA1090.shovill.skesa.bam
assembleAll.pl: Effective coverage for ./SneakerNet/assemblies/Philadelphia_CDC/Philadelphia_CDC.shovill.skesa.bam
assembleAll.pl: Effective coverage for ./SneakerNet/assemblies/LT2/LT2.shovill.skesa.bam
assembleAll.pl: Effective coverage for ./SneakerNet/assemblies/contaminated/contaminated.shovill.skesa.bam
assembleAll.pl: Metrics can be found in ./SneakerNet/forEmail/assemblyMetrics.tsv
sn_mlst.pl: Temporary directory is at /tmp/sn_mlst.pl.APv9me
sn_mlst.pl: Reading sample tsv at ./samples.tsv
sn_mlst.pl: MLST for contaminated
sn_mlst.pl: MLST for FA1090
sn_mlst.pl: MLST for 2010EL-1786
sn_mlst.pl: MLST for Philadelphia_CDC
sn_mlst.pl: MLST for LT2
sn_kraken.pl: Running Kraken on 2010EL-1786
sn_kraken.pl:   Database: /kraken-database
sn_kraken.pl: Running Kraken on contaminated
sn_kraken.pl:   Database: /kraken-database
sn_kraken.pl: Running Kraken on FA1090
sn_kraken.pl:   Database: /kraken-database
sn_kraken.pl: Running Kraken on LT2
sn_kraken.pl:   Database: /kraken-database
sn_kraken.pl: Running Kraken on Philadelphia_CDC
sn_kraken.pl:   Database: /kraken-database
sn_detectContamination-kraken.pl: Running Kraken on contaminated
sn_detectContamination-kraken.pl:   Database: /kraken-database
sn_detectContamination-kraken.pl: ./SneakerNet/kraken/contaminated
sn_detectContamination-kraken.pl: Including for email: ./SneakerNet/kraken/contaminated/report.html
sn_detectContamination-kraken.pl: Running Kraken on LT2
sn_detectContamination-kraken.pl:   Database: /kraken-database
sn_detectContamination-kraken.pl: ./SneakerNet/kraken/LT2
sn_detectContamination-kraken.pl: Including for email: ./SneakerNet/kraken/LT2/report.html
sn_detectContamination-kraken.pl: Running Kraken on FA1090
sn_detectContamination-kraken.pl:   Database: /kraken-database
sn_detectContamination-kraken.pl: ./SneakerNet/kraken/FA1090
sn_detectContamination-kraken.pl: Including for email: ./SneakerNet/kraken/FA1090/report.html
sn_detectContamination-kraken.pl: Running Kraken on 2010EL-1786
sn_detectContamination-kraken.pl:   Database: /kraken-database
sn_detectContamination-kraken.pl: ./SneakerNet/kraken/2010EL-1786
sn_detectContamination-kraken.pl: Including for email: ./SneakerNet/kraken/2010EL-1786/report.html
sn_detectContamination-kraken.pl: Running Kraken on Philadelphia_CDC
sn_detectContamination-kraken.pl:   Database: /kraken-database
sn_detectContamination-kraken.pl: ./SneakerNet/kraken/Philadelphia_CDC
sn_detectContamination-kraken.pl: Including for email: ./SneakerNet/kraken/Philadelphia_CDC/report.html
sn_detectContamination-mlst.pl: --mlstfasta was not given, but I set it anyway to /mlst-2.19.0/bin/../db/blast/mlst.fa
sn_detectContamination-mlst.pl: Filtering /mlst-2.19.0/bin/../db/blast/mlst.fa
sn_detectContamination-mlst.pl: Running colorid workflow
ERROR: while creating ./SneakerNet/colorid/PE.tsv: this file does not exist: /home/eriny/src/SneakerNet/t/M00123-18-001-test/FA1090_1.fastq.gz at /SneakerNet-0.14.0/scripts/../SneakerNet.plugins/sn_detectContamination-mlst.pl line 139.
SneakerNetPlugins.pl: ERROR with plugin sn_detectContamination-mlst.pl: ERROR running command
  /SneakerNet-0.14.0/scripts/../SneakerNet.plugins/sn_detectContamination-mlst.pl . --numcpus 8 at /SneakerNet-0.14.0/scripts/SneakerNetPlugins.pl line 99.

Use of uninitialized value $_[0] in substitution (s///) at /usr/share/perl/5.26/File/Basename.pm line 180.
fileparse(): need a valid pathname at /SneakerNet-0.14.0/scripts/SneakerNetPlugins.pl line 109.

```
