+++ SneakerNet.plugins/sn_sarscov2_assembleAll.pl
@@ -8,6 +8,7 @@ use Data::Dumper;
 use File::Basename qw/fileparse basename dirname/;
 use File::Temp qw/tempdir/;
 use File::Copy qw/mv cp/;
+use File::Spec;
 use Bio::SeqIO;
 use Bio::FeatureIO::gff;
 use List::Util qw/min max/;
@@ -34,6 +35,7 @@ sub main{
       #'run_prediction_metrics.pl (CG-Pipeline)'     => "echo CG Pipeline version unknown",
       'run_assembly_metrics.pl (CG-Pipeline)'       => "echo CG Pipeline version unknown",
       cat                             => 'cat --version | head -n 1',
+      # wget needed in SneakerNet.pm
       wget                            => 'wget --version | head -n 1',
       samtools                        => 'samtools 2>&1 | grep Version:',
       bcftools                        => 'bcftools 2>&1 | grep Version:',
@@ -41,7 +43,7 @@ sub main{
       seqtk                           => 'seqtk 2>&1 | grep -m 1 Version:',
       bgzip                           => 'bgzip --version | head -n1',
       tabix                           => 'tabix --version | head -n1',
-      'v-annotate.pl (VADR)'          => 'v-annotate.pl -h | grep -m 1 [0-9]'
+      #'v-annotate.pl (VADR)'          => 'v-annotate.pl -h | grep -m 1 [0-9]'
     }, $settings,
   );
 
@@ -82,7 +84,7 @@ sub assembleAll{
     logmsg "ASSEMBLE SAMPLE $sample";
 
     my $outdir="$dir/SneakerNet/assemblies/$sample";
-    my $outassembly="$outdir/$sample.bowtie2.bcftools.fasta";
+    my $outassembly=File::Spec->rel2abs("$outdir/$sample.bowtie2.bcftools.fasta");
     my $outgbk="$outdir/$sample.bowtie2.bcftools.gbk";
 
     # Run the assembly
@@ -111,46 +113,25 @@ sub assembleAll{
       mv("$asmMetricsFile.tmp", $asmMetricsFile) or die "ERROR: could not make $asmMetricsFile";
     }
 
-    # Genome annotation
-    #logmsg "PREDICT GENES FOR SAMPLE $sample";
-    my $vadrPass = "$outdir/vadr/vadr.vadr.pass.tbl";
-    my $vadrFail = "$outdir/vadr/vadr.vadr.fail.tbl";
-    if(!-e $vadrPass || !-e $vadrFail){
-      logmsg "ANNOTATE SAMPLE $sample";
-      my $tmpVadrDir=annotateFasta($sample,$outassembly,$settings);
-      mkdir("$outdir/vadr");
-      for(glob("$tmpVadrDir/*")){
-        next if(!-f $_);
-        cp($_, "$outdir/vadr/".basename($_)) or die "ERROR copying $_ to $outdir/vadr/: $!";
-      }
-    }
-    logmsg "Annotations can be found in $vadrPass and $vadrFail";
-
     my $cdsMetricsFile = "$outdir/predictionMetrics.tsv";
     if(!-e $cdsMetricsFile){
-      logmsg "Prediction metrics for $vadrPass";
-      my $passCds = readTbl($vadrPass, $settings);
-      #print Dumper $passCds;
-      logmsg "Prediction metrics for $vadrFail";
-      my $failCds = readTbl($vadrFail, $settings);
-      #print Dumper $failCds;
-
-      my $numPassed = $$passCds{numFeatures}{mat_peptide} || 0;
-      my $numFailed = $$failCds{numFeatures}{mat_peptide} || 0;
-
[…]
