### 2020-11-17 #2334 @lskatz
Hi @erinyoung, we were able to replicate the error.  Could you try the same thing but with these options?

```
--no email --no transfer --no save
```

We were able to avoid the error with our command like so inside of Docker
```
SneakerNetPlugins.pl --numcpus 16 --no email --no transfer --no save M00123-18-001-test/
```

It turns out that some of the functionality is difficult to get into the container and so we acknowledge that some of the plugins do not work as intended on the container interface.
