### 2020-11-18 ba96dbf
added threads to samtools view
