### 2020-11-18 4f08b21
updated Clints script
