### 2020-11-20 52f775f
added bowtie2 dependency for sarscov2 assembly
