### 2020-11-20 3d5b8b3
sars-cov-2 assembly documentation updated for algorithm and new columns
