### 2020-11-20 edf0168
SARS-CoV-2 (#56)
* supporting the _1 _2 format for fastq

* taxon properties for sars-cov-2

* emoticons displayed

* one eyebrow raised

* subbullet

* updated version

* added cds detection but there is no QC for it yet

* vadr annotation and QC

* some debugging statements

* more debugging

* more debugging INC

* find >dev/null

* push find command further down

* push find command further down

* fix find command to /opt/hostedtoolcache

* separately install Bio::* modules

* test bioperl

* cpanm to PWD

* added dependency packages listed on packages.ubuntu.com/bionic/bioperl

* apt-get
[…]
