+++ SneakerNet.plugins/sn_sarscov2_assembleAll.pl
@@ -1,0 +1,445 @@
+#!/usr/bin/env perl
+# Assemble genomes for SARS-CoV-2 amplicons
+
+use strict;
+use warnings;
+use Getopt::Long;
+use Data::Dumper;
+use File::Basename qw/fileparse basename dirname/;
+use File::Temp qw/tempdir/;
+use File::Copy qw/mv cp/;
+use File::Spec;
+use Bio::SeqIO;
+use Bio::FeatureIO::gff;
+use List::Util qw/min max/;
+
+use threads;
+use Thread::Queue;
+
+use FindBin qw/$RealBin/;
+use lib "$FindBin::RealBin/../lib/perl5";
+use SneakerNet qw/readTsv exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg fullPathToExec/;
+
+our $VERSION = "1.1";
+our $CITATION= "SARS-CoV-2 assembly plugin by Lee Katz.";
+
+local $0=fileparse $0;
+exit(main());
+
+sub main{
+  my $settings=readConfig();
+  GetOptions($settings,qw(version citation check-dependencies help tempdir=s debug numcpus=i force)) or die $!;
+  exitOnSomeSneakernetOptions({
+      _CITATION => $CITATION,
+      _VERSION  => $VERSION,
+      #'run_prediction_metrics.pl (CG-Pipeline)'     => "echo CG Pipeline version unknown",
+      'run_assembly_metrics.pl (CG-Pipeline)'       => "echo CG Pipeline version unknown",
+      cat                             => 'cat --version | head -n 1',
+      # wget needed in SneakerNet.pm
+      wget                            => 'wget --version | head -n 1',
+      samtools                        => 'samtools 2>&1 | grep Version:',
+      bcftools                        => 'bcftools 2>&1 | grep Version:',
+      cutadapt                        => 'cutadapt --version',
+      seqtk                           => 'seqtk 2>&1 | grep -m 1 Version:',
+      bgzip                           => 'bgzip --version | head -n1',
+      tabix                           => 'tabix --version | head -n1',
+      bowtie2                         => 'bowtie2 --version | grep -m 1 version',
+      'bowtie2-build'                   => 'bowtie2-build --version | grep -m 1 version',
+      #'v-annotate.pl (VADR)'          => 'v-annotate.pl -h | grep -m 1 [0-9]'
+    }, $settings,
+  );
+
+  die usage() if($$settings{help} || !@ARGV);
+  $$settings{VADRMODELDIR} ||= die "ERROR: VADRMODELDIR needs to be defined under config/settings.conf.
+  Please see for more help: https://github.com/nawrockie/vadr/wiki/Coronavirus-annotation#how-to-annotate-sars-cov-2-sequences-with-vadr-1";
+  $$settings{numcpus}||=1;
+  $$settings{tempdir}||=File::Temp::tempdir(basename($0).".XXXXXX",TMPDIR=>1,CLEANUP=>1);
+  logmsg "Temporary directory is at $$settings{tempdir}";
+
+  my $dir=$ARGV[0];
+  mkdir "$dir/SneakerNet/forEmail";
+
+  if($$settings{force}){
+    logmsg "--force was given; removing any existing assembly results.";
+    system("rm -rf $dir/SneakerNet/assemblies $dir/forEmail/assemblyMetrics.tsv");
+  }
+ 
+  mkdir "$dir/SneakerNet";
+  mkdir "$dir/SneakerNet/assemblies";
+
+  my $metricsOut=assembleAll($dir,$settings);
[…]
