### 2020-12-07 #2334 @erinyoung
Now I want to be a pain and run it on my data. I haven't been able to connect directly to my MiSeq. It seems like sneakernet can be run on fastq files and organize them appropriately. Here's what I tried:
```
docker run --rm -v $PWD:/data -v $KRAKENDB:/kraken-database -u $(id -u):$(id -g) lskatz/sneakernet:latest SneakerNet.roRun.pl --createsamplesheet -o M70330-20-001-test --numcpus 8 fastq --tempdir .
```
And got the following output
```
SneakerNet.roRun.pl: ERROR: no interop files were found in fastq
SneakerNet.roRun.pl: demultiplex sample sheets were not found. Creating one directly from fastq filenames into ./runData/samples.tsv
SneakerNet.roRun.pl: snok.txt not found. I will not read from it.
SneakerNet.roRun.pl: cp fastq/2020CK-00360-UT-M70330-200910_S8_L001_R1_001.fastq.gz to ./runData/2020CK-00360-UT-M70330-200910_S8_L001_R1_001.fastq.gz
ERROR: could not write to ./runData/2020CK-00360-UT-M70330-200910_S8_L001_R1_001.fastq.gz: Permission denied at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 207.
```
And the following files :
```
$ ls runData/
2020CK-00360-UT-M70330-200910_S8_L001_R1_001.fastq.gz  SampleSheet.csv  samples.tsv  snok.txt
```
It seems like it's trying to copy over my fastq R1 twice. `SampleSheet.csv`, `samples.tsv`, and `snok.txt` are empty.

Well...
```
$ cat runData/SampleSheet.csv
[Data]
```
