### 2020-12-11 #2334 @lskatz
This is cool @erinyoung !  We are happy it worked on the test data!  And it is really cool that you are trying it on your data.  @kapsakcj (our excellent containers expert on the team!) tried but could not replicate this error.  Although this issue is real, it unfortunately makes it so that we have to speculate on the solution.

I believe but do not know for sure that `./runData` might be pointing to a relative path in the container instead of from the root, ie, `/runData`.  Also with the weirdness of directory paths and container paths, I think but do not know for sure, avoiding `--tempdir` might help clear up another variable here.  One more thought though along those lines is that maybe Docker needs to have the tmpdir explicitly given?  For example: `docker run -v /tmp:/tmp ...` might be helpful.

And again, this is speculation and not likely, but just to help us remove any other variables in this issue, do you have read/write access to the input directory?

> It seems like it's trying to copy over my fastq R1 twice. SampleSheet.csv, samples.tsv, and snok.txt are empty.

I think but don't know for sure that what you're seeing in the log might be where it copies to a temp space and then back to the output folder.  This is as intended, because if a file is in the output directory then the pipeline knows that it is done with that file.  In other words, a file's presence in the output folder is a marker for "done."
