### 2020-12-11 #2334 @erinyoung
I do have read/write access, but I cannot copy over a file just to overwrite it. There needs to be a bit of time between the two.

I'll change the temp directory to see if that helps any.
