### 2021-01-12 60eeedc
added some error messaging for the HTML report
