### 2021-01-12 afddc07
more thoughtful assembly fail codes
