+++ SneakerNet.plugins/sn_passfail.pl
@@ -13,7 +13,7 @@ use List::Util qw/sum/;
 use lib "$FindBin::RealBin/../lib/perl5";
 use SneakerNet qw/readTsv @rankOrder %rankOrder readKrakenDir exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg/;
 
-our $VERSION = "6.0";
+our $VERSION = "6.1";
 our $CITATION="SneakerNet pass/fail by Lee Katz";
 
 $ENV{PATH}="$ENV{PATH}:/opt/cg_pipeline/scripts";
@@ -246,7 +246,7 @@ sub identifyBadRuns{
       # If we made it through all filters, then we have reached
       # the point where we can call it contaminated.
       $fail{kraken} = 1;
-      last; # no need to test any further because it has failed
+      last; # no need to test any further ranks because it has failed
     }
 
     # Did it pass by kraken / read classification
@@ -254,15 +254,74 @@ sub identifyBadRuns{
 
     #die Dumper [\%assemblyMetrics, {%{$assemblyMetrics{$samplename}}}, $samplename];
     my $minNs = $$sampleInfo{$samplename}{taxonRules}{assembly_percentN};
-    my $percentNs = $assemblyMetrics{$samplename}{percentNs};
-    if(!defined($percentNs)){
+    my $longestContigThreshold = $$sampleInfo{$samplename}{taxonRules}{longest_contig};
+    if(defined($minNs) && defined($longestContigThreshold)){
+      # Set up a subtest for the assembly.
+      # If any of these assembly metrics fail, then
+      # the assembly fails QC.
+      # If all are unknown, then the assembly QC is unknown.
+      # If all pass, then assembly QC passes.
+      # Definitions:
+      #   -1: unknown
+      #    0: passes
+      #    1: fails
+      my %failAssembly = (
+        minNs         => -1,
+        longestContig => -1,
+      );
+
+      # QC for minimum number of Ns allowed
+      my $percentNs = $assemblyMetrics{$samplename}{percentNs};
+      if(!defined($percentNs)){
+        $failAssembly{minNs} = -1;
+      }
+      elsif($percentNs > $minNs){
+        $failAssembly{minNs} = 1;
+      }
+      elsif($percentNs <= $minNs){
+        $failAssembly{minNs} = 0;
+      }
+
+      # QC for longest contig
+      my $longestContig = $assemblyMetrics{$samplename}{longestContiguous};
+      if(!defined($longestContig)){
+        $failAssembly{longestContig} = -1;
+      }
+      elsif($longestContig >= $longestContigThreshold){
+        $failAssembly{longestContig} = 0;
+      }
+      elsif($longestContig <  $longestContigThreshold){
+        $failAssembly{longestContig} = 1;
+      }
+
+      # Compile assembly failure codes into a single code
+      my $numFailed = 0;
+      my $numTests  = 0;
+      while(my($why, $failCode) = each(%failAssembly)){
+        if($failCode==1){
+          $numFailed++;
+        }
+        elsif($failCode==-1){
+          next;
+        }
+        $numTests++;
+      }
+      if($numTests < 1){
+        $fail{assembly} = -1;
+      }
+      else{
+        if($numFailed > 0){
+          $fail{assembly} = 1;
+        } else {
+          $fail{assembly} = 0;
+        }
+      }
+
+    }
+    # If assembly thresholds are not defined, then it is
+    # unknown whether this sample passed.
+    else {
       $fail{assembly} = -1;
-    }
-    elsif($percentNs > $minNs){
-      $fail{assembly} = 1;
-    }
-    elsif($percentNs <= $minNs){
-      $fail{assembly} = 0;
     }
 
     # Save the culmination of pass/fail for this sample
