### 2021-01-12 6011ca1
new assembly thresholds
