### 2021-01-13 3727348
updated assembly QC docs
