+++ docs/plugins/sn_passfail.pl.md
@@ -17,7 +17,9 @@ read quality, from the read metrics plugin.
 Table with columns
 
 * sample
-* assembly - pass or fail based on number of Ns (threshold defined in `taxonProperties.conf`) or number of genes
+* assembly - if all sub quality checks pass, then this passes. If all are unknown, then this is unknown. If any fail, then this fails.
+  * number of Ns (threshold defined in `taxonProperties.conf`) or number of genes
+  * longest contiguous contig - longest tract of genome between two Ns
 * coverage - pass or fail
 * quality  - pass or fail
 * kraken - pass or fail 
