### 2021-01-13 #2334 @erinyoung
```
docker run --rm -v $PWD:/data -v $KRAKENDB:/kraken-database -v /tmp:/tmp -u $(id -u):$(id -g) lskatz/sneakernet:latest SneakerNet.roRun.pl --createsamplesheet -o M70330-20-001-test --numcpus 8 fastq
```

Still gives me the same error.

```
ERROR: could not write to /tmp/SneakerNet.roRun.pl.KvbPif/runData/2020CK-00360-UT-M70330-200910_S8_L001_R1_001.fastq.gz: Permission denied at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 207.
```

This time there is no file. I thought that maybe the directory `/tmp/SneakerNet.roRun.pl.KvbPif/runData` might need to be created first, so I tried that and still got the same error.

BTW, this is a different error than before. Before the file would get copied over, now there is no file getting copied over.
