### 2021-01-16 b6ceed8
mapping reads to ref and then in-house trimming of primers before assembly
