+++ SneakerNet.plugins/sn_sarscov2_assembleAll.pl
@@ -6,7 +6,7 @@ use warnings;
 use Getopt::Long;
 use Data::Dumper;
 use File::Basename qw/fileparse basename dirname/;
-use File::Temp qw/tempdir/;
+use File::Temp qw/tempdir tempfile/;
 use File::Copy qw/mv cp/;
 use File::Spec;
 use Bio::SeqIO;
@@ -203,16 +203,28 @@ sub assembleAll{
 sub assembleSample{
   my($sample,$sampleInfo,$settings)=@_;
 
-  my($R1,$R2) = @{ $$sampleInfo{fastq} };
+  my($R1orig,$R2orig) = @{ $$sampleInfo{fastq} };
 
-  if(!$R1 || !-e $R1){
+  if(!$R1orig || !-e $R1orig){
     logmsg "Could not find R1 for $sample. Skipping.";
     return "";
   }
-  if(!$R2 || !-e $R2){
+  if(!$R2orig || !-e $R2orig){
     logmsg "Could not find R2 for $sample. Skipping.";
     return "";
   }
+
+  # Get the reference ready
+  my $ref = $$sampleInfo{taxonRules}{reference_fasta};
+  # Check for bowtie2 index files
+  if(!-e "$ref.1.bt2"){
+    my $log = "$ref.bowtie2-build.log";
+    logmsg "Index not found. bowtie2-build log can be found at $log";
+    command("bowtie2-build $ref $ref 2> $log");
+  }
+
+  # Filter to reads that map to the reference
+  my($R1,$R2) = readsThatMapTo($ref, $R1orig, $R2orig, $settings);
 
   # Dealing with a small file size
   # TODO look at read metrics instead
@@ -231,16 +243,11 @@ sub assembleSample{
   system("rm -rf '$outdir'"); # make sure any previous runs are gone
   mkdir $outdir;
 
-  my $ref = $$sampleInfo{taxonRules}{reference_fasta};
-
-  # Check for bowtie2 index files
-  if(!-e "$ref.1.bt2"){
-    my $log = "$ref.bowtie2-build.log";
-    logmsg "Index not found. bowtie2-build log can be found at $log";
-    command("bowtie2-build $ref $ref 2> $log");
+  my $primersBed = $$sampleInfo{taxonRules}{primers_bed};
+  if(!defined($primersBed)){
+    die "ERROR: the primers_bed file is not defined. This is usually derived from the primers_bed_url key found in taxonProperties.conf.";
   }
-
-  ($R1, $R2) = adapterTrim($R1, $R2, $settings);
+  ($R1, $R2) = adapterTrim($R1, $R2, $primersBed, $ref, $settings);
 
   logmsg "Mapping reads from $sample to $ref";
   my $samtoolsThreads = $$settings{numcpus}-1;
@@ -273,6 +280,30 @@ sub assembleSample{
   unlink("$outdir/out.vcf.gz.tbi");
 
   return $outdir
+}
+
+# Filter to just reads that map to a reference
+# The magic behind this is mapping and then samtools -F 4 -F 8
+#   => read not unmapped and mate not unmapped
+sub readsThatMapTo{
+  my($ref, $R1orig, $R2orig, $settings) = @_;
+
+  my $workingDir = File::Temp::tempdir("readsThatMapTo.XXXXXX", DIR=>$$settings{tempdir});
+
+  my $samtoolsThreads = $$settings{numcpus}-1;
+  my $sam = "$workingDir/bowtie2.sam";
[…]
