### 2021-01-19 11254ef
trimmomatic and also plot coverage
