+++ SneakerNet.plugins/sn_sarscov2_assembleAll.pl
@@ -19,8 +19,10 @@ use Thread::Queue;
 use FindBin qw/$RealBin/;
 use lib "$FindBin::RealBin/../lib/perl5";
 use SneakerNet qw/readTsv exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg fullPathToExec/;
+use GD; # for fonts
+use GD::Graph::lines;
 
-our $VERSION = "1.2";
+our $VERSION = "1.3";
 our $CITATION= "SARS-CoV-2 assembly plugin by Lee Katz.";
 
 # A message to show in the report if any
@@ -42,7 +44,7 @@ sub main{
       wget                            => 'wget --version | head -n 1',
       samtools                        => 'samtools 2>&1 | grep Version:',
       bcftools                        => 'bcftools 2>&1 | grep Version:',
-      cutadapt                        => 'cutadapt --version',
+      'trimmomatic'                   => 'trimmomatic -version 2>&1 | grep -v _JAVA',
       seqtk                           => 'seqtk 2>&1 | grep -m 1 Version:',
       bgzip                           => 'bgzip --version | head -n1',
       tabix                           => 'tabix --version | head -n1',
@@ -243,6 +245,7 @@ sub assembleSample{
   system("rm -rf '$outdir'"); # make sure any previous runs are gone
   mkdir $outdir;
 
+  logmsg "Trimming adapters";
   my $primersBed = $$sampleInfo{taxonRules}{primers_bed};
   if(!defined($primersBed)){
     die "ERROR: the primers_bed file is not defined. This is usually derived from the primers_bed_url key found in taxonProperties.conf.";
@@ -278,6 +281,36 @@ sub assembleSample{
   unlink($sam);
   unlink("$outdir/out.vcf.gz");
   unlink("$outdir/out.vcf.gz.tbi");
+
+  # make a graph
+  my $png = "$outdir/plot.png";
+  my @depthOfContig = ();
+  open(my $fh, "samtools depth -aa $sortedBam |") or die "ERROR: could not run samtools depth on $sortedBam: $!";
+  while(<$fh>){
+    chomp;
+    my($contig, $pos, $depth) = split /\t/;
+    $depthOfContig[$pos] = $depth;
+  }
+  close $fh;
+
+  my $graph = new GD::Graph::lines();
+  $graph->set(
+    title  => "Coverage of $sample",
+    x_label => "pos",
+    y_label => "depth",
+  );
+
+  $graph->set_legend_font(GD::Font->Small);
+
+  my $gd = $graph->plot([
+    [1..scalar(@depthOfContig)],
+    \@depthOfContig,
+  ]);
+  
+  open(my $pngFh, ">", $png) or die "ERROR: could not write to $png: $!";
+  binmode($pngFh);
+  print $pngFh $gd->png;
+  close $pngFh;
 
   return $outdir
 }
@@ -393,7 +426,8 @@ sub adapterTrim{
   my($R1in, $R2in, $primersBed, $ref, $settings) = @_;
   my $threads = $$settings{numcpus};
 
-  # Need to get sequences from reference with bed coordinates
+  ## Need to get sequences from reference with bed coordinates
+  # get reference contigs
   my $seqin = Bio::SeqIO->new(-file=>$ref);
   my %seq;
   while(my $seq=$seqin->next_seq){
@@ -401,8 +435,11 @@ sub adapterTrim{
   }
[…]
