### 2021-01-20 #2334 @lskatz
@lpantano, @lfaller, @erinyoung

Thanks all for participating in the review of our manuscript. The reviewer comments are insightful and have helped strengthen our software significantly. Over time, we addressed many documentation issues which made the software much better; we have also addressed the multiple dependency issues by creating a container; and after the container became unwieldy, we spent significant time creating a multistage build container (something we had never done before).

We were wondering if we could propose a debugging session with Dr. Young as we are unable to replicate her current error, and we believe some kind of interactive session could help clear things up more quickly / facilitate the review process.  We’re excited about what we’ve built with your terrific feedback, and because of the length of time to uncover and resolve these software challenges we thought it might help us move forward with getting to a point where this work could be published. Let us know what you think of this proposal, or if there is another way you would like to approach this.
