### 2021-01-22 #2334 @lpantano
Thanks @erinyoung, I totally understand. Is that the last comment? because if it is, i will try to find help to see if anybody else has this issue and if it is like that, work to fix it so it doesn't need to go through you. let me know.thanks!
