### 2021-01-22 #2334 @lpantano
I can help with debugging as well. can you try this @erinyoung :

I am assuming you can write in the folder you are running the command, I tried this one in my home directory, and I have a folder `fastq` in the same place you are running the command. Notice the path in the command in docker starts with `/data/`, that would be the same folder as where you are running the command because of this: `$PWD:/data` parameter in docker. so `data` inside docker command is the same than `PWD` in your computer.

```
docker run --rm -v $PWD:/data -v $KRAKENDB:/kraken-database -v /tmp:/tmp -u $(id -u):$(id -g) lskatz/sneakernet:latest SneakerNet.roRun.pl --createsamplesheet -o /data/M70330-20-001-test --numcpus 8 /data/fastq
```
