### 2021-01-26 #2334 @kapsakcj
Hi @erinyoung , I will demo the steps to analyzing a MiSeq run using the same docker image that you are using for SneakerNet.
I have tried to replicate your computing environment as best as I can. Hopefully this will help you run SneakerNet on your own data, but @lskatz  and I are available for a troubleshooting and screensharing session if necessary. Please let us know if you have questions or if we can clarify anything.

As @lskatz mentioned previously, you may need to mount your local `/tmp` directory to `/tmp` inside the container with `docker run -v /tmp:/tmp`. Ensure you have sufficient space in your local `/tmp` directory (hard to say exactly how much is needed, perhaps 1GB minimum?) and if not, please mount another location where you have sufficient space to `/tmp` inside the container. Also ensure that you have read and write permissions on your `$PWD`, `/tmp`, and your input directories and files.

Please do not use the `--tempdir` option with the SneakerNet scripts, as we believe it may cause more issues, especially while running SneakerNet inside a container.

Ensure we are using the same exact docker image for SneakerNet by checking the imageID (third column below)
NOTE: the image ID on your computer might be different the one shown here, see next step
```
$ docker images | grep sneakernet
lskatz/sneakernet                           latest                        567f79c9dce6   3 months ago    3.31GB
```

Verify that the docker image is the same version that is available under dockerhub at lskatz/sneakernet:latest.
This sha256 hash (7c1cc…) matches the hash listed on [dockerhub for lskatz/sneakernet:latest](https://hub.docker.com/r/lskatz/sneakernet/tags?page=2&ordering=last_updated)
```
$ docker image inspect 567f79c9dce6 | grep -A 1 RepoDigests
        "RepoDigests": [
            "lskatz/sneakernet@sha256:7c1cc91f2e6ae5154f85f4ec2439d2ccc9aa0cf50e8112782de80f7c4b9fedc8"
```

Verify that the docker image contains SneakerNet 0.14.0
```
$ docker run --rm lskatz/sneakernet:latest SneakerNetPlugins.pl --version
SneakerNet 0.14.0
```

Kraken1 database location and contents. Any kraken1 database will do, pending that it was created properly.
```
$ ls /home/pjx8/downloads/kraken1-db/minikraken_20171019_8GB
database.idx  database.kdb  MiniKraken.README  taxonomy

# set KRAKENDB variable to this path
$ export KRAKENDB=/home/pjx8/downloads/kraken1-db/minikraken_20171019_8GB
```

Below is my PWD for this demo; it contains `M347-20-010/` which is the output directory from a MiSeq sequencing run which contains the fastq.gz files.
(just for explanation, your MiSeq output dir can be named anything) The name of the directory represents the sequencing run name, which contains `<MiSeqID>-<year>-<run number>` (Your run name will obviously be different, this is specific to our lab)
```
$ ls -F
M347-20-010/
 ```

Listing contents, just to show there is WGS data for two samples here, corresponding with two Vibrio isolates.
Your input directory should look somewhat similar to this, but will differ depending on the number of samples on the sequencing run
```
$ ll M347-20-010/
total 336052
drwxrwx--- 3 pjx8 pjx8      4096 Jan 25 14:13 ./
drwxrwxr-x 3 pjx8 pjx8      4096 Jan 25 14:12 ../
-rwxrwx--- 1 pjx8 pjx8 103299728 Jan 25 14:13 2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz*
-rwxrwx--- 1 pjx8 pjx8 115358223 Jan 25 14:12 2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz*
-rwxrwx--- 1 pjx8 pjx8  56676995 Jan 25 14:13 2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz*
-rwxrwx--- 1 pjx8 pjx8  63791709 Jan 25 14:12 2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz*
-rwxrwx--- 1 pjx8 pjx8      3558 Jan 25 14:13 config.xml*
drwxrwx--- 3 pjx8 pjx8      4096 Jan 25 14:13 QC/
-rwxrwx--- 1 pjx8 pjx8       697 Jan 25 14:13 SampleSheet.csv*
-rwxrwx--- 1 pjx8 pjx8   2351730 Jan 25 14:13 Undetermined_S0_L001_R1_001.fastq.gz*
-rwxrwx--- 1 pjx8 pjx8   2600740 Jan 25 14:13 Undetermined_S0_L001_R2_001.fastq.gz*
```

First, run the `SneakerNet.roRun.pl` script to prepare and format the Sequencing run directory for SneakerNet analysis. This will create an output directory that is the formatted for SneakerNet and has copies of your input files.
NOTE: It’s important to specify the input and output PATHs both begin with `/data`
```
# broken into two lines for readability
$ docker run -v $PWD:/data -v $KRAKENDB:/kraken-database -u $(id -u):$(id -g) lskatz/sneakernet:latest \
    SneakerNet.roRun.pl /data/M347-20-010 -o /data/M347-20-010-SN-formatted --numcpus 8
```
This ran to completion with warnings, but no errors. Log here:

```
SneakerNet.roRun.pl: snok.txt not found. I will not read from it.
SneakerNet.roRun.pl: cp /data/M347-20-010/2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz
SneakerNet.roRun.pl: cp /data/M347-20-010/2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz
SneakerNet.roRun.pl: cp /data/M347-20-010/2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz
SneakerNet.roRun.pl: cp /data/M347-20-010/2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz
SneakerNet.roRun.pl: cp /data/M347-20-010/Undetermined_S0_L001_R1_001.fastq.gz to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/Undetermined_S0_L001_R1_001.fastq.gz
SneakerNet.roRun.pl: cp /data/M347-20-010/Undetermined_S0_L001_R2_001.fastq.gz to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/Undetermined_S0_L001_R2_001.fastq.gz
SneakerNet.roRun.pl: cp /data/M347-20-010/SampleSheet.csv to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/SampleSheet.csv
SneakerNet.roRun.pl: cp /data/M347-20-010/Data/Intensities/BaseCalls/config.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/config.xml
WARNING: could not copy /data/M347-20-010/Data/Intensities/BaseCalls/config.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/config.xml: No such file or directory
  Making a blank file instead. at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 205.
SneakerNet.roRun.pl: cp /data/M347-20-010/CompletedJobInfo.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/CompletedJobInfo.xml
WARNING: could not copy /data/M347-20-010/CompletedJobInfo.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/CompletedJobInfo.xml: No such file or directory
  Making a blank file instead. at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 205.
SneakerNet.roRun.pl: cp /data/M347-20-010/runParameters.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/runParameters.xml
WARNING: could not copy /data/M347-20-010/runParameters.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/runParameters.xml: No such file or directory
  Making a blank file instead. at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 205.
SneakerNet.roRun.pl: cp /data/M347-20-010/GenerateFASTQRunStatistics.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/GenerateFASTQRunStatistics.xml
WARNING: could not copy /data/M347-20-010/GenerateFASTQRunStatistics.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/GenerateFASTQRunStatistics.xml: No such file or directory
  Making a blank file instead. at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 205.
SneakerNet.roRun.pl: cp /data/M347-20-010/RunInfo.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/RunInfo.xml
WARNING: could not copy /data/M347-20-010/RunInfo.xml to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/RunInfo.xml: No such file or directory
  Making a blank file instead. at /SneakerNet-0.14.0/scripts/SneakerNet.roRun.pl line 205.
SneakerNet.roRun.pl: cp /data/M347-20-010/QC/InterOp/ControlMetricsOut.bin to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/ControlMetricsOut.bin
SneakerNet.roRun.pl: cp /data/M347-20-010/QC/InterOp/CorrectedIntMetricsOut.bin to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/CorrectedIntMetricsOut.bin
SneakerNet.roRun.pl: cp /data/M347-20-010/QC/InterOp/ExtractionMetricsOut.bin to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/ExtractionMetricsOut.bin
SneakerNet.roRun.pl: cp /data/M347-20-010/QC/InterOp/IndexMetricsOut.bin to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/IndexMetricsOut.bin
SneakerNet.roRun.pl: cp /data/M347-20-010/QC/InterOp/QMetricsOut.bin to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/QMetricsOut.bin
SneakerNet.roRun.pl: cp /data/M347-20-010/QC/InterOp/TileMetricsOut.bin to /tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/TileMetricsOut.bin
created directory '/data/M347-20-010-SN-formatted'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/snok.txt' -> '/data/M347-20-010-SN-formatted/snok.txt'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz' -> '/data/M347-20-010-SN-formatted/2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz' -> '/data/M347-20-010-SN-formatted/2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz' -> '/data/M347-20-010-SN-formatted/2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz' -> '/data/M347-20-010-SN-formatted/2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/Undetermined_S0_L001_R1_001.fastq.gz' -> '/data/M347-20-010-SN-formatted/Undetermined_S0_L001_R1_001.fastq.gz'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/Undetermined_S0_L001_R2_001.fastq.gz' -> '/data/M347-20-010-SN-formatted/Undetermined_S0_L001_R2_001.fastq.gz'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/SampleSheet.csv' -> '/data/M347-20-010-SN-formatted/SampleSheet.csv'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/config.xml' -> '/data/M347-20-010-SN-formatted/config.xml'
created directory '/data/M347-20-010-SN-formatted/QC'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/CompletedJobInfo.xml' -> '/data/M347-20-010-SN-formatted/QC/CompletedJobInfo.xml'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/runParameters.xml' -> '/data/M347-20-010-SN-formatted/QC/runParameters.xml'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/GenerateFASTQRunStatistics.xml' -> '/data/M347-20-010-SN-formatted/QC/GenerateFASTQRunStatistics.xml'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/RunInfo.xml' -> '/data/M347-20-010-SN-formatted/QC/RunInfo.xml'
created directory '/data/M347-20-010-SN-formatted/QC/InterOp'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/ControlMetricsOut.bin' -> '/data/M347-20-010-SN-formatted/QC/InterOp/ControlMetricsOut.bin'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/CorrectedIntMetricsOut.bin' -> '/data/M347-20-010-SN-formatted/QC/InterOp/CorrectedIntMetricsOut.bin'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/ExtractionMetricsOut.bin' -> '/data/M347-20-010-SN-formatted/QC/InterOp/ExtractionMetricsOut.bin'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/IndexMetricsOut.bin' -> '/data/M347-20-010-SN-formatted/QC/InterOp/IndexMetricsOut.bin'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/QMetricsOut.bin' -> '/data/M347-20-010-SN-formatted/QC/InterOp/QMetricsOut.bin'
copied '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/TileMetricsOut.bin' -> '/data/M347-20-010-SN-formatted/QC/InterOp/TileMetricsOut.bin'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/Undetermined_S0_L001_R1_001.fastq.gz'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/SampleSheet.csv'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/CompletedJobInfo.xml'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/RunInfo.xml'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/runParameters.xml'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/ControlMetricsOut.bin'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/TileMetricsOut.bin'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/ExtractionMetricsOut.bin'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/CorrectedIntMetricsOut.bin'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/QMetricsOut.bin'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp/IndexMetricsOut.bin'
removed directory '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/InterOp'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC/GenerateFASTQRunStatistics.xml'
removed directory '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/QC'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/snok.txt'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/config.xml'
removed '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData/Undetermined_S0_L001_R2_001.fastq.gz'
removed directory '/tmp/SneakerNet.roRun.pl.jd1zSZ/runData'
rmdir /tmp/SneakerNet.roRun.pl.jd1zSZ
```

Now run default SneakerNet workflow on SN-formatted directory using `SneakerNetPlugins.pl` script.
NOTE: ensure to include the options `--no email --no transfer --no save` to prevent errors.
NOTE: you do not need to specify an output directory with this command, as it will create an sub-directory (`M347-20-010-SN-formatted/SneakerNet/`) with the resulting files.
```
# broken into two lines for readability
$ docker run -v $PWD:/data -v $KRAKENDB:/kraken-database -u $(id -u):$(id -g) lskatz/sneakernet:latest \
    SneakerNetPlugins.pl --numcpus 8 --no email --no transfer --no save /data/M347-20-010-SN-formatted
```
Log here:

```
sn_parseSampleSheet.pl: Wrote samples file to ./samples.tsv
addReadMetrics.pl: Tempdir is /tmp/addReadMetrics.pl.SJD2Lo
addReadMetrics.pl: Reading sample ./SampleSheet.csv
addReadMetrics.pl: Running fast read metrics
addReadMetrics.pl: read metrics for ./2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz
addReadMetrics.pl: read metrics for ./2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz
addReadMetrics.pl: read metrics for ./2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz
addReadMetrics.pl: read metrics for ./2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz
addReadMetrics.pl: read metrics for ./Undetermined_S0_L001_R1_001.fastq.gz
addReadMetrics.pl: read metrics for ./Undetermined_S0_L001_R2_001.fastq.gz
addReadMetrics.pl: Backfilling values in ./readMetrics.tsv
addReadMetrics.pl: Decided that ./2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz is Vibrio with expected genome size 4000000. Calculated coverage: 17.37
addReadMetrics.pl: Decided that ./2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz is Vibrio with expected genome size 4000000. Calculated coverage: 17.39
addReadMetrics.pl: Decided that ./2016V-1195-M347-20-010_S1_L001_R1_001.fastq.gz is Vibrio with expected genome size 4000000. Calculated coverage: 31.73
addReadMetrics.pl: Decided that ./2016V-1195-M347-20-010_S1_L001_R2_001.fastq.gz is Vibrio with expected genome size 4000000. Calculated coverage: 31.83
assembleAll.pl: Temporary directory is at /tmp/assembleAll.pl.DnWUD3
assembleAll.pl: ASSEMBLE SAMPLE 2017V-1003-M347-20-010
assembleAll.pl: Assembling 2017V-1003-M347-20-010
[shovill] Hello stranger
[shovill] You ran: /shovill/shovill-1.1.0/bin/shovill --outdir /tmp/assembleAll.pl.DnWUD3/2017V-1003-M347-20-010 --R1 /data/M347-20-010-SN-formatted/2017V-1003-M347-20-010_S2_L001_R1_001.fastq.gz --R2 /data/M347-20-010-SN-formatted/2017V-1003-M347-20-010_S2_L001_R2_001.fastq.gz --assembler skesa --cpus 8 --keepfiles
[shovill] This is shovill 1.1.0
[shovill] Written by Torsten Seemann
[shovill] Homepage is https://github.com/tseemann/shovill
[shovill] Operating system is linux
[shovill] Perl version is v5.26.1
[shovill] Machine has 24 CPU cores and 125.85 GB RAM
[shovill] Using bwa - /usr/local/bin/bwa | Version: 0.7.17-r1188
[shovill] Using flash - /usr/local/bin/flash | FLASH v1.2.11
[shovill] Using java - /usr/bin/java | openjdk version "11.0.8" 2020-07-14
[shovill] Using kmc - /usr/local/bin/kmc | K-Mer Counter (KMC) ver. 3.1.1 (2019-05-19)
[shovill] Using lighter - /usr/local/bin/lighter | Lighter v1.1.1
[shovill] Using megahit - /usr/local/bin/megahit | MEGAHIT v1.1.3
[shovill] Using megahit_toolkit - /usr/local/bin/megahit_toolkit | v1.1.3
[shovill] Using pigz - /usr/bin/pigz | pigz 2.4
[shovill] Using pilon - /pilon/pilon | Pilon version 1.22 Wed Mar 15 16:38:30 2017 -0400
[shovill] Using samclip - /usr/local/bin/samclip | samclip 0.4.0
[shovill] Using samtools - /usr/local/bin/samtools | Version: 1.10 (using htslib 1.10)
[shovill] Using seqtk - /usr/local/bin/seqtk | Version: 1.3-r106
[shovill] Using skesa - /usr/local/bin/skesa | SKESA 2.4.0
[shovill] Using spades.py - /SPAdes-3.14.1-Linux/bin/spades.py | SPAdes genome assembler v3.14.1
[shovill] Using trimmomatic - /trimmomatic/trimmomatic | 0.38
[shovill] Using velvetg - /usr/local/bin/velvetg | Version 1.2.10
[shovill] Using velveth - /usr/local/bin/velveth | Version 1.2.10
[shovill] Found spades version: 003014000
[shovill] Using tempdir: /tmp/rXijsz_Dro
[shovill] Changing into folder: /tmp/assembleAll.pl.DnWUD3/2017V-1003-M347-20-010
[shovill] Collecting raw read statistics with 'seqtk'
[shovill] Running: seqtk fqchk -q3 \/data\/M347\-20\-010\-SN\-formatted\/2017V\-1003\-M347\-20\-010_S2_L001_R1_001\.fastq\.gz >/tmp/owOgRSVn7g 2>&1 | sed 's/^/[seqtk] /' | tee -a shovill.log
[shovill] Read stats: total_bp = 139008764
[shovill] Read stats: max_len = 151
[shovill] Read stats: min_len = 35
[shovill] Read stats: avg_len = 147
[shovill] Estimating genome size by counting unqiue 21-mers > frequency 10
[shovill] Running: kmc -sm -m8 -t8 -k21 -ci10 \/data\/M347\-20\-010\-SN\-formatted\/2017V\-1003\-M347\-20\-010_S2_L001_R1_001\.fastq\.gz /tmp/VYfMWx3kBx/kmc /tmp/VYfMWx3kBx 2>&1 | sed 's/^/[kmc] /' | tee -a shovill.log
[shovill] Using genome size 3537630 bp
[shovill] Estimated sequencing depth: 39 x
[shovill] No read depth reduction requested or necessary.
[shovill] Appending -Xmx16g to _JAVA_OPTIONS
[shovill] Running: ln -sf \/data\/M347\-20\-010\-SN\-formatted\/2017V\-1003\-M347\-20\-010_S2_L001_R1_001\.fastq\.gz R1.fq.gz 2>&1 | sed 's/^/[ln] /' | tee -a shovill.log
[shovill] Running: ln -sf \/data\/M347\-20\-010\-SN\-formatted\/2017V\-1003\-M347\-20\-010_S2_L001_R2_001\.fastq\.gz R2.fq.gz 2>&1 | sed 's/^/[ln] /' | tee -a shovill.log
[shovill] Average read length looks like 147 bp
[shovill] Automatically setting --minlen to 73
[shovill] Setting k-mer range to (31 .. 110)
[shovill] Estimated K-mers: 31 51 71 91 [kn=5, ks=20, kmin=31, kmax=110]
[shovill] Using kmers: 31,51,71,91
[shovill] Correcting reads with 'Lighter'
[shovill] Running: lighter -od . -r R1.fq.gz -r R2.fq.gz -K 32 3537630 -t 8 -maxcor 1 2>&1 | sed 's/^/[lighter] /' | tee -a shovill.log
[shovill] Overlapping/stitching PE reads with 'FLASH'
[shovill] Running: flash -m 20 -M 151 -d . -o flash -z -t 8 R1.cor.fq.gz R2.cor.fq.gz 2>&1 | sed 's/^//' | tee -a shovill.log
[shovill] Assembling reads with 'skesa'
[shovill] Running: skesa --gz  --fastq flash.extendedFrags.fastq.gz --fastq flash.notCombined_1.fastq.gz,flash.notCombined_2.fastq.gz --use_paired_ends --contigs_out skesa.fasta --min_contig 1 --memory 16 --cores 8 --vector_percent 1 2>&1 | sed 's/^/[skesa] /' | tee -a shovill.log
[shovill] Checking for assembly errors in skesa.fasta
[shovill] Running: bwa index skesa.fasta 2>&1 | sed 's/^/[bwa-index] /' | tee -a shovill.log
[shovill] Running: samtools faidx skesa.fasta 2>&1 | sed 's/^/[faidx] /' | tee -a shovill.log
[…]
