### 2021-01-26 9dc3e25
added image path to properties.txt
