### 2021-01-27 20253e3
compress log msg; better error message if no alleles found in any genomes
