+++ .github/workflows/unittests.yml
@@ -163,7 +163,8 @@ jobs:
       - name: apt-get install
         run:  |
           sudo apt-get update
-          sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl cpanminus
+          sudo apt-get -y install libgd-dev
+          sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl cpanminus 
           sudo apt-get -y install build-essential bioperl-run ncbi-blast+ prodigal sendmail python3 python3-pip
           # shovill things
           sudo apt-get -y install bwa velvet wget pigz libncurses5-dev libbz2-dev liblzma-dev libfindbin-libs-perl trimmomatic
