### 2021-02-01 #2334 @lpantano
Hi @lskatz, thank you for the detailed explanation. I have made this work with some other files as well. @erinyoung, if this is ok with you, I think this is functional, and it may need some docker knowledge or sys admin to install it easily in some systems. Are you ok with publishing this since other people can run a real miseq data finally? if so, can you check the last item in the list under functionality if you agree with the statements, other wise, feel free to comment on any other issue. Thank you a lot!
