### 2021-02-01 #2334 @lpantano
@lskatz , last question. I did get an error during this step `SneakerNet.roRun.pl` if FASTQ files were labeled as `_1/2` for the paired files, and it worked well when I renamed to `_R1/2`. Is this a known requirements? I will mention that in the quick start if it is. If this should work, I am happy to send the commands to reproduce it. thanks!
