### 2021-02-03 #2334 @erinyoung
I didn't realize I needed a kraken1 database. I've changed that.
```
$ ls -F
00_env.t                  07_sn_passfail.t                     34_mlst.t                     53_sarscov2_assembly.t          M00123-20-001-sarscov2/
01_sn_parseSampleSheet.t  08_transferFilesToRemoteComputers.t  36_staramr.t                  90_universalOptionsInPlugins.t  report.html
02_addReadMetrics.t       09_emailWhoever.t                    40_metagenomics.t             91_pluginsDocumentation.t
03_assembleAll.t          10_mlstRaw.t                         42_readMetricsMetagenomics.t  futureTests/
05_kraken.t               30_assemblyWorkflow_env.t            44_krakenMetagenomics.t       M00123-18-001-test/
06_baseBalance.t          31_sn_assemblyWorkflow_init.t        51_sarscov2_readMetrics.t     M00123-18-002-test/
```
