### 2021-02-04 #2334 @lskatz
> @lskatz , last question. I did get an error during this step `SneakerNet.roRun.pl` if FASTQ files were labeled as `_1/2` for the paired files, and it worked well when I renamed to `_R1/2`. Is this a known requirements? I will mention that in the quick start if it is. If this should work, I am happy to send the commands to reproduce it. thanks!

Good catch @lpantano .  We documented that idiosyncrasy here https://github.com/lskatz/SneakerNet/commit/d8a2f3f10f379ad1d65e8df34b19cea437f79aa8 and here https://github.com/lskatz/SneakerNet/commit/8cfab72600da1d86cbc16c9ff72128ee1b6cb51f, and also in the CLI usage in the script itself here https://github.com/lskatz/SneakerNet/commit/41ae36dee8c881ebe8feff19ce0a724e9b314667.
