### 2021-02-04 #2334 @lskatz
@erinyoung I am a little perplexed still when I look at how you are able to write to `/tmp/` with

    $  docker run -v $PWD:/data -v $KRAKENDB:/kraken-database -v /tmp:/tmp -u $(id -u):$(id -g) lskatz/sneakernet:latest echo "test" >  /tmp/whatever.txt

but cannot write to tmp with the script.

Could we try to prepare your data the alternate way, described in our documentation here?  https://github.com/lskatz/SneakerNet/blob/master/docs/SneakerNetInput.md#building-from-scratch
