### 2021-02-05 #2334 @lpantano
@erinyoung, are those files:

```
2020CK-00360-UT-M70330-200910_S8_L001_R1_001.fastq.gz
2020CK-00360-UT-M70330-200910_S8_L001_R2_001.fastq.gz
```

 in your computer or in a mounted folder from another file system?

@lskatz, can you explain what these lines will do:

```
  my $return = link($from, $to) ||
    File::Copy::cp($from,$to) or warn "WARNING: could not copy $from to $to: $!\n  Making a blank file instead.";
  open(my $fh, ">>", $to) or die "ERROR: could not write to $to: $!";
  close $fh;
```

She doesn't see the warning, so in theory it is copying the file or making a link (maybe?) but then it fails when it is doing the `open` part?  it could be that is doing a `link` (hard) but then it is giving an error when trying to open to write? If you could get me through those lines I can think of other ways to reproduce. I don't really know PERL, so not sure if I understand each step is happening there.

I have tried everything, not giving permission to the file to read, having the file to be a broken symlink, but always it finishes (with warnings), but finishes.

I am looking for another reviewer who can do this quick, if @erinyoung cannot finish this without making sure the tool run in that system. I know she is busy with COVID19 research, so thank you for your time, and sorry this is being problematic at this time. I appreciate authors' time to try to help with this, I know it has been a while this is open. I'll do my best to find the best solution for everybody. Thanks
