### 2021-02-09 #2334 @lskatz
> She doesn't see the warning, so in theory it is copying the file or making a link (maybe?) but then it fails when it is doing the `open` part? it could be that is doing a `link` (hard) but then it is giving an error when trying to open to write? If you could get me through those lines I can think of other ways to reproduce. I don't really know PERL, so not sure if I understand each step is happening there.

How interesting!  I think you uncovered the bug @lpantano.  This is part of a larger function where I overwrote `cp`.  This bug has never been triggered before and I think I know why it is triggering an error finally.

This line

      my $return = link($from, $to) ||
        File::Copy::cp($from,$to) or warn "WARNING: could not copy $from to $to: $!\n  Making a blank file instead.";

makes a hard link and if that is not possible (`||`), then it runs the actual `cp` from perl.  With so many large files, a hard link has been much faster and advantageous.  In @erinyoung 's test case, the hard link must have worked.  However right after that, the file itself must not be writeable and so this line

    open(my $fh, ">>", $to) or die "ERROR: could not write to $to: $!";
    close $fh;

ends in error even though nothing is actually written to it.  So somehow, the source file is not writeable and is being hard linked.

In response, I have made a much more careful file check here and it does not create a zero byte file unless the target file does not exist at all: https://github.com/lskatz/SneakerNet/commit/a663a894f564a015d4bb05b47bd8b12a60e57917#diff-495115971f2da9ad2f2aa888784b67455c22ea947094507390a5a0459f64e9b6
