### 2021-02-09 #2334 @lpantano
oh, thanks! will we get this fix with the latest docker image so @erinyoung may try if she finds the time?
