### 2021-02-09 0b4840c
bumped the version; bumped the Dockerfile version
