### 2021-02-10 4586031
adding apt-get update before libgd install
