### 2021-02-10 #2334 @lskatz
@lpantano Thanks for pointing that out.  I have updated our Docker image from v0.14 of SneakerNet to 0.19.6 to include all recent changes.  Although most changes thus far are nonfunctional since v0.14 (ie documentation), this is an important fix to address the latest issue.  This image can be found in `lskatz/sneakernet:edge` or more permanently `lskatz/sneakernet:sha-d7255a84`.
