### 2021-02-11 #2334 @lpantano
@erinyoung if you have time to repeat that command but using: `lskatz/sneakernet:edge` instead of `lskatz/sneakernet:latest`, let us know. thanks!  @druvus, I would use this latest image when you test, just in case you face the same error. Thanks both
