### 2021-03-18 #2334 @druvus
@lpantano sorry for not pinging you.  I have used the Docker edge image to run through the test data and an internal MiSeq run. It seems to work for me using Docker but I haven't tried manual installation so far. It is a lot of dependencies but it looks like would be quite possible, especially in combination with conda.
