### 2021-03-24 25c194c
curl instead of wget for minikraken
