+++ .github/workflows/unittests.yml
@@ -163,7 +163,7 @@ jobs:
       - name: apt-get install
         run:  |
           sudo apt-get update
-          sudo apt-get -y install libgd-dev
+          sudo apt-get -y install libgd-dev openssl-1.1.1g
           sudo apt-get -y install tree libberkeleydb-perl libdb-dev bioperl perl cpanminus curl 
           sudo apt-get -y install build-essential bioperl-run ncbi-blast+ prodigal sendmail python3 python3-pip
           # shovill things
