### 2021-03-29 0a1eda1
fix certs according to rec from Curtis
https://github.com/lskatz/SneakerNet/commit/8131f8ceaacb31534ee8ffb67e7faa4ab2f4218f#commitcomment-48843653
