+++ .github/workflows/unittests.yml
@@ -201,8 +201,8 @@ jobs:
         run: |
           sudo mkdir -v /kraken-database
           cd /kraken-database
-          #sudo wget --no-check-certificate https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz
-          curl https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz --output minikraken_20171019_4GB.tgz
+          sudo wget --secure-protocol=TLSv1 --no-check-certificate https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz
+          # sudo curl https://ccb.jhu.edu/software/kraken/dl/minikraken_20171019_4GB.tgz --output minikraken_20171019_4GB.tgz
           sudo tar -zxf minikraken_20171019_4GB.tgz 
           sudo mv -nv minikraken_20171013_4GB/* . # flatten db dir
           sudo ls -lhSR /kraken-database
