### 2021-03-31 0961a82
build my own crappy kraken database
