### 2021-03-31 7e8549d
making a small local kraken database
