### 2021-03-31 6d2fcb9
chmod kraken db
