### 2021-03-31 350d644
adding back the downloading of minikraken
