### 2021-03-31 cb12517
remove custom small kraken database
