### 2021-04-11 #2334 @lskatz
Hi @lpantano, sorry this last week was really busy and every day I tried to address this.  I spent some time making sure that the unit testing works because something broke on the GitHub Actions VM in the certificate authority.  I think I just have to make sure Zenodo and the DOI is correct and then I can check it all off.  I should be able to do this before Thursday of this week, hopefully much sooner.

Thank you also @druvus for your time in reviewing our submission.
