### 2021-04-14 #2334 @kyleniemeyer
Hi @lskatz @lpantano, I'm the AEIC on duty this week, just looking to help get this past the finish line!

Everything looks good, except for one minor issue in the paper: it is missing the Statement of Need section, which is a [required element](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of JOSS papers.

Could you add this section? You can likely move some of your existing text, rather than write something completely new. Once that is done, we will be ready to accept!
